BRCAPRO

Runs PanelPRO with genes = c("BRCA1", "BRCA2") and cancers = c("Breast", "Ovarian", "Contralateral"), as in PanelPRO:::MODELPARAMS$BRCAPRO.

BRCAPRO(...)

Arguments

...

Additional arguments to be passed into PanelPRO.

See also

Examples

data(test_fam_1)
BRCAPRO(test_fam_1, parallel = FALSE)
#> Contralateral cancer(s) not in pedigree, so they will be removed from the model specification.
#> Your model has 2 cancers - Breast, Ovarian and 2 genes - BRCA1_hetero_anyPV, BRCA2_hetero_anyPV.
#> Setting max.mut to 2.
#> ID 6,29 has tumor marker testing results but is unaffected for relevant cancer, so testing will be ignored.
#> Germline testing results for BRCA1 are assumed to be for default variant BRCA1_hetero_anyPV.
#> Germline testing results for BRCA2 are assumed to be for default variant BRCA2_hetero_anyPV.
#> ID 3 's Ancestry has been changed to nonAJ to meet hereditary consistency.
#> ID 10 's race has been changed to All_Races to meet hereditary consistency.
#> ID 9 's Ancestry has been changed to nonAJ to meet hereditary consistency.
#> ID 16,17 's Ancestry has been changed to nonAJ to meet hereditary consistency.
#> ID 29,30 's Ancestry has been changed to nonAJ to meet hereditary consistency.
#> ID 33,34 's race has been changed to All_Races to meet hereditary consistency.
#> ID 33,34 's Ancestry has been changed to nonAJ to meet hereditary consistency.
#> Loading required package: foreach
#> Loading required package: rngtools
#> ID 6's lower and upper bounds for the probability of carrying any pathogenic variant are (0.30599, 0.30604).
#> The value ranges in the results are driven by the different imputed missing ages at each imputation, so please consider including more age information when possible to reduce the range widths.
#> $posterior.prob
#> $posterior.prob$`6`
#>                                   genes     estimate        lower        upper
#> 1                            noncarrier 0.6939891572 0.6939579966 0.6940081987
#> 2                    BRCA1_hetero_anyPV 0.3048909722 0.3048772649 0.3048993558
#> 3                    BRCA2_hetero_anyPV 0.0007889223 0.0007696017 0.0008205312
#> 4 BRCA1_hetero_anyPV.BRCA2_hetero_anyPV 0.0003309483 0.0003228438 0.0003442073
#> 
#> 
#> $future.risk
#> $future.risk$`6`
#> $future.risk$`6`$Breast
#>   ByAge   estimate      lower      upper
#> 1    60 0.04727254 0.04726917 0.04727805
#> 2    65 0.09413804 0.09413203 0.09414788
#> 3    70 0.13827732 0.13826933 0.13829037
#> 4    75 0.17626900 0.17625959 0.17628439
#> 5    80 0.20681051 0.20680003 0.20682766
#> 6    85 0.22907973 0.22906848 0.22909815
#> 7    90 0.24161181 0.24160015 0.24163090
#> 
#> $future.risk$`6`$Ovarian
#>   ByAge   estimate      lower      upper
#> 1    60 0.03602836 0.03602727 0.03603013
#> 2    65 0.07192665 0.07192434 0.07193043
#> 3    70 0.10474092 0.10473738 0.10474671
#> 4    75 0.13251405 0.13250940 0.13252165
#> 5    80 0.15535805 0.15535254 0.15536708
#> 6    85 0.17200390 0.17199784 0.17201381
#> 7    90 0.17978899 0.17978271 0.17979926
#> 
#> 
#>