Contralateral Breast Cancer Penetrances

Get a subject's age specific crude penetrances of contralateral breast cancer, based on eight subject characteristics.

CBCnoncarrierCrude(
  diag.age,
  bc1.type,
  fdr.status,
  ER.status,
  birth1.age,
  antiEst,
  HighRiskP,
  BD,
  bc1.t.size,
  sex,
  race,
  had1stBC = 1
)

Arguments

diag.age

numeric value, age of 1st breast cancer diagnosis

bc1.type

numeric value, type of first breast cancer, "Invasive" or Unk = 3, "Invasive_DCIS" = 2, "DCIS" = 1 ("DCIS" not used in PanelPRO)

fdr.status

numeric value, if first degree relatives have history of breast cancer, Yes = 1, No = 2, Unk = 3 (Unk not used in PanelPRO)

ER.status

numeric value, ER test result, Unk = 3, Pos = 2, Neg = 1

birth1.age

numeric value, age at first child birth, <30/nulliparous = 1, 30-39 = 2, 40+ = 3, Unk = 4

antiEst

numeric value, indicates if anti-estrogen therapy was used to treat the first breast cancer, Yes = 1, No = 2, Unk = 3

HighRiskP

numeric value, indicates history of high risk preneoplasia, Yes = 1, No/Unk = 2

BD

numeric value based on BI-RADS breast density ratings "a" to "d". The categories differ for non-Hispanic Black women versus other women. For women who are NOT non-Hispanic Black: "a" or entirely fat = 4, "b" or scattered = 3, "c" or heterogeneously dense = 2, "d" or extremely dense = 1, Unk = 5. For women who ARE non-Hispanic Black: "a" (entirely fat) or "b" (scattered) are 1, "c" (heterogeneously dense) or "d" (extremely dense) are 2, Unk is 3.

bc1.t.size

number representing a category of 1st BC tumor size, 1 for sizes T0/T1/T2, 2 for T3/T4, 3 for Tis, or 4 for unknown size.

sex

string, subject's sex, one of c("Male", "Female"), default is "Female". If sex is not "Female", function will return an all 0's vector because the cbcrisk model is not valid for males.

race

number, 2 if non-Hispanic Black, 1 otherwise.

had1stBC

numeric value, indicates if subject previously diagnosed with primary breast cancer, Yes = 1, No = 0. If had1stBC is not 1 then then function will return an all 0's vector because the model relies on a known diagnosis age of primary breast cancer.

Value

A named vector of penetrances from age 1 to age PanelPRO:::MAXAGE

where the vector's names are the ages. Penetrances are 0 from age 1

through the age of first breast cancer diagnosis. Penetrances are all 0 if had1stBC is not 1.

Details

Uses the cbcrisk function from the cbcrisk package for female CBC penetrances and subsets the an array of CBC noncarrier penetrances from ASK2ME for males. The inputs of this function are designed to be compatible with the outputs of the getCBCParams function.