MMRPRO

Runs PanelPRO with genes = c("MLH1", "MSH2", "MSH6") and cancers = c("Colorectal", "Endometrial"), as in PanelPRO:::MODELPARAMS$MMRPRO.

MMRPRO(...)

Arguments

...

Additional arguments to be passed into PanelPRO.

See also

Examples

data(test_fam_3)
MMRPRO(test_fam_3, parallel = FALSE)
#> Your model has 2 cancers - Colorectal, Endometrial and 3 genes - MLH1_hetero_anyPV, MSH2_hetero_anyPV, MSH6_hetero_anyPV.
#> Setting max.mut to 1.
#> Riskmods that are NA will be removed
#> ID 12's lower and upper bounds for the probability of carrying any pathogenic variant are (0.00076, 0.00103).
#> The value ranges in the results are driven by the different imputed missing ages at each imputation, so please consider including more age information when possible to reduce the range widths.
#> $posterior.prob
#> $posterior.prob$`12`
#>               genes     estimate        lower        upper
#> 1        noncarrier 0.9991500638 0.9989676964 0.9992405827
#> 2 MLH1_hetero_anyPV 0.0002791342 0.0002552200 0.0003305956
#> 3 MSH2_hetero_anyPV 0.0002249832 0.0002079940 0.0002639838
#> 4 MSH6_hetero_anyPV 0.0003458189 0.0002962034 0.0004377241
#> 
#> 
#> $future.risk
#> $future.risk$`12`
#> $future.risk$`12`$Colorectal
#>    ByAge    estimate       lower       upper
#> 1     49 0.001703109 0.001699884 0.001710138
#> 2     54 0.004744383 0.004737735 0.004758829
#> 3     59 0.008370528 0.008360301 0.008392661
#> 4     64 0.012848289 0.012834473 0.012878039
#> 5     69 0.018257421 0.018240165 0.018294374
#> 6     74 0.024270956 0.024250541 0.024314415
#> 7     79 0.030627571 0.030604543 0.030676340
#> 8     84 0.036416917 0.036392122 0.036469249
#> 9     89 0.040518905 0.040493272 0.040572916
#> 10    94 0.042462643 0.042436872 0.042516932
#> 
#> $future.risk$`12`$Endometrial
#>    ByAge estimate lower upper
#> 1     49        0     0     0
#> 2     54        0     0     0
#> 3     59        0     0     0
#> 4     64        0     0     0
#> 5     69        0     0     0
#> 6     74        0     0     0
#> 7     79        0     0     0
#> 8     84        0     0     0
#> 9     89        0     0     0
#> 10    94        0     0     0
#> 
#> 
#>