PanelPRO master function

This is the main function of the PanelPRO package. It combines the pre-processing steps (checking the user-supplied pedigree, imputing missing ages, and setting up the database of model parameters) with the calculation of the genotypic probability distributions and future risks of cancer for the specified probands.

PanelPRO(
  pedigree,
  model_spec = "PanPRO22",
  proband = pedigree$ID[pedigree$isProband == 1],
  genes = NULL,
  cancers = NULL,
  database = PanelPRODatabase,
  unknown.race = UNKNOWN_RACE,
  unknown.ancestry = UNKNOWN_ANCESTRY,
  impute.missing.ages = TRUE,
  allow.intervention = TRUE,
  ignore.proband.germ = FALSE,
  max.mut = NULL,
  remove.miss.cancers = TRUE,
  iterations = 20,
  max.iter.tries = NULL,
  parallel = TRUE,
  debug = FALSE,
  net = FALSE,
  age.by = 5,
  random.seed = 42L,
  plusBCRAT = FALSE,
  bcrat.vars = NULL,
  rr.bcrat = NULL,
  rr.pop = rr.ref
)

Arguments

pedigree

A data frame of family history information with the following columns. Unknown or missing values should be explicitly coded as NA.

ID: A numeric value; ID for each individual. There should not be any duplicated entries.

Sex: A numeric value; 0 for female and 1 for male. Missing entries are not currently supported.

MotherID: A numeric value; unique ID for someone's mother.

FatherID: A numeric value; unique ID for someone's father.

isProband: A numeric value; 1 if someone is a proband, 0 otherwise. This will be overridden by the proband argument in PanelPRO, if it is specified. At least one proband should be specified by either the isProband column or proband. Multiple probands are supported.

CurAge: A numeric value; the age of censoring (current age if the person is alive or age of death if the person is dead). Ages ranging from 1 to 94 are allowed.

isAffX: A numeric value; the affection status of cancer X, where X is a short cancer code (see Details). Affection status should be encoded as 1 if the individual was diagnosed, 0 otherwise. Missing entries are not currently supported.

AgeX: A numeric value; the age of diagnosis for cancer X, where X is a short cancer code (see Details). Ages ranging from 1 to 94 are allowed. If the individual was not diagnosed for a given cancer, their affection age should be encoded as NA and will be ignored otherwise.

isDead: A numeric value; 1 if someone is dead, 0 otherwise. Missing entries are assumed to be 0.

race: A character string; expected values are "All_Races", "AIAN" (American Indian and Alaska Native), "Asian", "Black", "White", "Hispanic", "WH" (white Hispanic), and "WNH" (non-white Hispanic) (see PanelPRO:::RACE_TYPES). Asian-Pacific Islanders should be encoded as "Asian". Race information will be used to select the cancer and death by other causes penetrances used in the model. Missing entries are recoded as the unknown.race argument, which defaults to PanelPRO:::UNKNOWN_RACE.

Ancestry: A character string; expected values are "AJ"(Ashkenazi Jewish), "nonAJ", and "Italian" (see PanelPRO:::ANCESTRY_TYPES). The ancestry information is used to determine the allele frequencies of BRCA1 and BRCA2, if they are included in the model. Missing entries are recoded as the unknown.ancestry argument, which defaults to PanelPRO:::UNKNOWN_ANCESTRY.

Twins: A numeric value; 0 for non-identical/single births, 1 for the first set of identical twins/multiple births in the family, 2 for the second set, etc. Missing entries are assumed to be 0.

riskmod: A character list; expected values are "Mastectomy", "Hysterectomy", and "Oophorectomy" (see data.frame(PanelPRO:::RISKMODS)). These preventative interventions will be used to modify the cancer penetrances for breast, endometrial, and ovarian cancer, respectively. Currently, "Mastectomy" refers to a bilateral prophylactic mastectomy (occurring prior to first breast cancer).

interAge: A numeric list; the age of intervention for each risk modifier in riskmod. For example, riskmod = list("Mastectomy", "Hysterectomy") and interAge = {45, 60} indicates that the individual had a mastectomy at age 45 and a hysterectomy at age 60.

There can be optional columns for germline testing results (e.g. BRCA1, MLH1) or tumor marker testing results. ER, PR, CK14, CK5.6 and HER2 are tumor markers associated with breast cancer that will modify the likelihood of phenotypes associated with BRCA1 and BRCA2. MSI is a tumor marker for colorectal cancer that will modify the likelihoods associated with MLH1, MSH2 MSH6 and PMS2. For each of these optional columns, positive results should be coded as 1, negative results should be coded as 0, and unknown results should be coded as NA.

model_spec

One of the following character strings indicating a pre-specified model: "BRCAPRO", "BRCAPRO5", "BRCAPRO6", "MMRPRO", "PanPRO11", and "PanPRO22". See PanelPRO:::MODELPARAMS for more information. Either model_spec or genes and cancers should be specified by the user. When both input types are supplied, the input from genes and cancers will take precedence. The default is "PanPRO22", the largest pre-specified model currently supported by the package.

proband

A numeric value or vector of the unique IDs in pedigree corresponding to the proband(s) for whom carrier probabilities and future risk predictions should be estimated. The default is pedigree$ID[pedigree$isProband == 1], in which case probands will be inferred from the isProband column in pedigree, if it exists and is valid. Otherwise, the IDs in proband will override the contents of pedigree$isProband. The user can also specify proband = "All", in which case all IDs in the ped will be treated as probands.

genes

A character vector of the genes of interest. The default is NULL. Available options are listed in PanelPRO:::GENE_TYPES. Either model_spec or genes and cancers should be specified by the user. When both input types are supplied, the input from genes and cancers will take precedence. If at least one cancer is specified in cancers, but no genes are specified in genes, this will be treated as a valid model that calculates future risk predictions based on the model parameters in database, without incorporating the family history in pedigree, and does not return any carrier probabilities.

cancers

A character vector of the cancers of interest. Available options are listed in PanelPRO:::CANCER_TYPES. The default is NULL. Either model_spec or genes and cancers should be specified by the user. When both input types are supplied, the input from genes and cancers will take precedence. If at least one gene is specified in genes, but no cancers are specified in cancers, this will be treated as a valid model that calculates carrier probability estimates based on the model parameters in database, without incorporating the family history in pedigree, and does not return any future risks.

database

A structured list providing the parameters of the model. The default is PanelPRODatabase. Users who wish to provide their own model parameters should supply a list with the same structure as PanelPRODatabase.

unknown.race

A character string indicating the default race to use when pedigree$race is missing or unsupported. The default is PanelPRO:::UNKNOWN_RACE.

unknown.ancestry

A character string indicating the default ancestry to use when pedigree$Ancestry is missing or unsupported. The default is PanelPRO:::UNKNOWN_ANCESTRY.

impute.missing.ages

A logical value indicating whether missing ages should be multiply imputed. The default is TRUE. If set to FALSE, an error will be raised if any individuals have missing current or cancer affection ages.

allow.intervention

A logical value indicating if cancer penetrance risk modifiers should be considered. The default is TRUE. Currently supported interventions are preventative mastectomies, hysterectomies, and oophorectomies (see data.frame(PanelPRO:::RISKMODS)).

ignore.proband.germ

A logical value indicating if the proband(s)'s germline testing results (if provided in ped) should be considered. The default is FALSE, in which case the proband(s)'s germline testing results will not be ignored.

max.mut

The maximum number of simultaneous mutations allowed. The default is NULL, in which case a maximum of 2 simultaneous mutations will be used when pedigree has 30 or fewer members, and 1 otherwise.

remove.miss.cancers

A logical value indicating whether cancers in cancers or implied by model_spec should be removed from the model specification when they are missing in pedigree. The default is TRUE.

iterations

A numeric value indicating the number of iterations that should be run when multiply imputing missing ages. The default is 20.

max.iter.tries

A numeric value indicating the maximum number of iterations that should be attempted when multiply imputing missing ages. (Invalid iterations typically occur when an individual is missing both their current and cancer ages. Current ages for living individuals are imputed first; if a valid cancer affection cannot be found, subject to the upper bound set by the imputed current age, this imputation run will be discarded.) The default is NULL, in which case the maximum number of tries will be set to five times iterations.

parallel

A logical value indicating whether the peeling-paring algorithm should be run in parallel when missing ages are imputed multiple times. The default is TRUE.

debug

A logical value indicating whether to debug using browsers and sequential imputation. The default is FALSE. Setting debug = TRUE is only useful for package developers.

net

A logical value indicating whether to return net or crude future risk estimates. The default is FALSE, in which case crude future risk estimates will be returned. Net risk is the probability of developing a given cancer by the specified age, in a world where there is no death from other causes. Crude risk is the probability of developing a given cancer by the specified age and not having died from other causes beforehand.

age.by

The fixed age interval for the future risk calculation. The default is 5, in which case if the proband's current age is 30, risk predictions will be calculated at ages 35, 40, 45, etc.

random.seed

The random seed (a numeric value) to set for imputing missing ages. The default is 42L.

plusBCRAT

A logical value indicating whether to use the BRCAPRO+BCRAT model. The default is FALSE. If plusBCRAT = TRUE, the user should specify one of bcrat.vars or rr.bcrat. The BCRAPRO+BCRAT model is only run if the user also sets net = FALSE and breast cancer is included in the model specification.

bcrat.vars

A data frame of BCRAT covariates, where each row corresponds to a proband in pedigree. The expected columns are:

ID: Numeric ID for each proband. Should not contain duplicated entries.

Age: Initial age (should be >=20 and <90).

NBiops: Number of benign breast biopsies (99 if unknown).

Hyp: A logical value indicating whether the individual has had a breast biopsy with atypical hyperplasia (0 if no, 1 if yes, 99 if unknown).

AgeMen: Age at menarche (99 if unknown)

AgeFLB: Age at first live birth (98 if nulliparous, 99 if unknown).

NumRel: The number of female first-degree relatives with breast cancer (99 if unknown).

Race: A character string indicating race ("White", "Black", "Hispanic", "FHispanic", "Asian", or "Other"; "Hispanic" corresponds to U.S.-born Hispanic while "FHispanic" corresponds to foreign-born Hispanic).

The default is NULL. bcrat.vars is only used if plusBCRAT = TRUE, net = FALSE, breast cancer is included in the model specification, and the user does not input rr.bcrat. The user only needs to input one of rr.bcrat and bcrat.vars to run BCRAPRO+BCRAT; if both are specified, rr.bcrat will be used for the model.

rr.bcrat

A data frame with 3 columns: the IDs of the probands (ID), the probands' BCRAT relative risks for age < 50 (rr1), and the probands' BCRAT relative risks for age >= 50 (rr2). When net = FALSE and breast cancer is included in the model, these values will modify the future risk calculation. The default is NULL. rr.bcrat is only used if plusBCRAT = TRUE, net = FALSE, and breast cancer is included in the model specification. The user only needs to input one of rr.bcrat and bcrat.vars to run BCRAPRO+BCRAT; if both are specified, rr.bcrat will be used for the model.

rr.pop

A data frame of race-specific for estimates of 1 / (1 - population attributable risk). The default is rr.ref, which was estimated from NHIS 2015.

Value

A list with two components:

posterior.prob: A list where each element corresponds to a proband in proband and consists of a data frame of posterior carrier probabilities. There are three columns: estimate (the estimate if no ages were imputed, or the estimate averaged over all age imputations), lower (NA if no ages were imputed, or the minimum result for all age imputations), and upper (NA if no ages were imputed, or the maximum result for all age imputations). The rows correspond to the different genotypes. If no genes were in the model specification, each proband's data frame of carrier probabilities is replaced by the message "No carrier probabilities were requested by the model specification." If ages were imputed, ImputeRange messages giving the lower and upper bounds for the probability of carrying any pathogenic variant will be printed for each proband. A ZeroProb warning is raised if one or more of the estimated probabilities is 0 (or 1), germline testing results were incorporated, and default sensitivities/specifities of 1 were used.

future.risk: A list where each element corresponds to a proband in proband and consists of a list of data frames of future risk estimates at fixed intervals for each cancer in the model. Each data frame has three columns: estimate (the estimate if no ages were imputed, or the estimate averaged over all age imputations), lower (NA if no ages were imputed, or the minimum result for all age imputations), and upper (NA if no ages were imputed, or the maximum result for all age imputations). Each row corresponds to the risk estimate for at a different age, e.g. 35, 40, 45, etc. when age.by = 5 for a proband whose current age is 30. If no cancers were in the model specification, each proband's list of future risk estimates is replaced by the message "No future risk estimates were requested by the model specification."

Details

The PanelPRO master function combines pre-processing of model parameters and the pedigree with calculation of the posterior carrier probabilities and future risk estimates. First, a model-specific database is built by subsetting database for the genes and cancers implied by model_spec or genes and cancers (see buildDatabase). Then, pedigree is checked for having a valid, consistent structure, and missing ages are imputed (see checkFam). If pedigree fails a check, corrections will be made when possible, and an informative error message will be issued if not. Finally, the carrier probabilities and future risk predictions for the user-specified probands are estimated and returned (see PanelPROCalc).

For convenience, the PanelPRO package exports wrapper functions for all of the pre-specified models: BRCAPRO, BRCAPRO5, BRCAPRO6, PanelPRO11, and PanelPRO22.

To represent a cancer in pedigree, we need to use short cancer codes. See data.frame(PanelPRO:::CANCER_NAME_MAP) for available options. For example, BC is the short name that maps to Breast, so breast cancer affection status and age of diagnosis should be recorded in columns named isAffBC and AgeBC.

When the corresponding cancer penetrance information is available in PanelPRODatabase, PanelPRO calculates and outputs carrier probabilities for both homozygous and heterozygous mutations of the genes included in model. MUTYH is the only gene for which we currently have both homozygous and heterozygous penetrances; homozygous mutations for all other genes are considered non-viable (see PanelPRO:::ALL_GENE_VARIANT_TYPES).

Sets of identical twins/multiple births only contribute to the likelihood once; see .twinsLikMod for more information. Carrier probabilities returned for individuals in the same set will be identical.

The R package cbcrisk, which is on GitHub but is not on CRAN, will be automatically installed when the function runs, if is is not installed already.

See also

Examples

PanelPRO(test_fam_1, model_spec = "BRCAPRO", proband = 5, parallel = FALSE)
#> Contralateral cancer(s) not in pedigree, so they will be removed from the model specification.
#> Your model has 2 cancers - Breast, Ovarian and 2 genes - BRCA1_hetero_anyPV, BRCA2_hetero_anyPV.
#> Setting max.mut to 2.
#> `proband` argument does not match with "isProband" column in the pedigree, using `proband`
#> ID 6,29 has tumor marker testing results but is unaffected for relevant cancer, so testing will be ignored.
#> Germline testing results for BRCA1 are assumed to be for default variant BRCA1_hetero_anyPV.
#> Germline testing results for BRCA2 are assumed to be for default variant BRCA2_hetero_anyPV.
#> ID 3 's Ancestry has been changed to nonAJ to meet hereditary consistency.
#> ID 10 's race has been changed to All_Races to meet hereditary consistency.
#> ID 9 's Ancestry has been changed to nonAJ to meet hereditary consistency.
#> ID 16,17 's Ancestry has been changed to nonAJ to meet hereditary consistency.
#> ID 29,30 's Ancestry has been changed to nonAJ to meet hereditary consistency.
#> ID 33,34 's race has been changed to All_Races to meet hereditary consistency.
#> ID 33,34 's Ancestry has been changed to nonAJ to meet hereditary consistency.
#> ID 5's lower and upper bounds for the probability of carrying any pathogenic variant are (1, 1).
#> The value ranges in the results are driven by the different imputed missing ages at each imputation, so please consider including more age information when possible to reduce the range widths.
#> Warning: Some of the estimated probabilities are 0 or 1, most likely due to the incorporation of reported germline testing results (by default, we assume that the sensitivity and specificity for all germline tests is 1). However, this does not guarantee that the person is a noncarrier or mutation carrier; a more accurate estimate may be achieved with user-specified sensitivities and specificities.
#> $posterior.prob
#> $posterior.prob$`5`
#>                                   genes   estimate      lower       upper
#> 1                            noncarrier 0.00000000 0.00000000 0.000000000
#> 2                    BRCA1_hetero_anyPV 0.99867676 0.99857092 0.998725730
#> 3                    BRCA2_hetero_anyPV 0.00000000 0.00000000 0.000000000
#> 4 BRCA1_hetero_anyPV.BRCA2_hetero_anyPV 0.00132324 0.00127427 0.001429079
#> 
#> 
#> $future.risk
#> $future.risk$`5`
#> $future.risk$`5`$Breast
#>   ByAge     estimate        lower        upper
#> 1    70 0.0002896075 0.0002880944 0.0002928779
#> 2    75 0.0006028101 0.0005997432 0.0006094383
#> 3    80 0.0009191193 0.0009145951 0.0009288971
#> 4    85 0.0011888034 0.0011830966 0.0012011371
#> 5    90 0.0013368604 0.0013304304 0.0013507573
#> 
#> $future.risk$`5`$Ovarian
#>   ByAge estimate lower upper
#> 1    70        0     0     0
#> 2    75        0     0     0
#> 3    80        0     0     0
#> 4    85        0     0     0
#> 5    90        0     0     0
#> 
#> 
#> 
PanelPRO(test_fam_1, genes = c("BRCA1", "BRCA2"), 
         cancers = c("Breast", "Ovarian"), proband = 5, parallel = FALSE)
#> Your model has 2 cancers - Breast, Ovarian and 2 genes - BRCA1_hetero_anyPV, BRCA2_hetero_anyPV.
#> Setting max.mut to 2.
#> `proband` argument does not match with "isProband" column in the pedigree, using `proband`
#> ID 6,29 has tumor marker testing results but is unaffected for relevant cancer, so testing will be ignored.
#> Germline testing results for BRCA1 are assumed to be for default variant BRCA1_hetero_anyPV.
#> Germline testing results for BRCA2 are assumed to be for default variant BRCA2_hetero_anyPV.
#> ID 3 's Ancestry has been changed to nonAJ to meet hereditary consistency.
#> ID 10 's race has been changed to All_Races to meet hereditary consistency.
#> ID 9 's Ancestry has been changed to nonAJ to meet hereditary consistency.
#> ID 16,17 's Ancestry has been changed to nonAJ to meet hereditary consistency.
#> ID 29,30 's Ancestry has been changed to nonAJ to meet hereditary consistency.
#> ID 33,34 's race has been changed to All_Races to meet hereditary consistency.
#> ID 33,34 's Ancestry has been changed to nonAJ to meet hereditary consistency.
#> ID 5's lower and upper bounds for the probability of carrying any pathogenic variant are (1, 1).
#> The value ranges in the results are driven by the different imputed missing ages at each imputation, so please consider including more age information when possible to reduce the range widths.
#> Warning: Some of the estimated probabilities are 0 or 1, most likely due to the incorporation of reported germline testing results (by default, we assume that the sensitivity and specificity for all germline tests is 1). However, this does not guarantee that the person is a noncarrier or mutation carrier; a more accurate estimate may be achieved with user-specified sensitivities and specificities.
#> $posterior.prob
#> $posterior.prob$`5`
#>                                   genes   estimate      lower       upper
#> 1                            noncarrier 0.00000000 0.00000000 0.000000000
#> 2                    BRCA1_hetero_anyPV 0.99867676 0.99857092 0.998725730
#> 3                    BRCA2_hetero_anyPV 0.00000000 0.00000000 0.000000000
#> 4 BRCA1_hetero_anyPV.BRCA2_hetero_anyPV 0.00132324 0.00127427 0.001429079
#> 
#> 
#> $future.risk
#> $future.risk$`5`
#> $future.risk$`5`$Breast
#>   ByAge     estimate        lower        upper
#> 1    70 0.0002896075 0.0002880944 0.0002928779
#> 2    75 0.0006028101 0.0005997432 0.0006094383
#> 3    80 0.0009191193 0.0009145951 0.0009288971
#> 4    85 0.0011888034 0.0011830966 0.0012011371
#> 5    90 0.0013368604 0.0013304304 0.0013507573
#> 
#> $future.risk$`5`$Ovarian
#>   ByAge estimate lower upper
#> 1    70        0     0     0
#> 2    75        0     0     0
#> 3    80        0     0     0
#> 4    85        0     0     0
#> 5    90        0     0     0
#> 
#> 
#> 

output <- PanelPRO22(test_fam_2, parallel = FALSE, max.mut = 1)
#> Brain, Breast, Colorectal, Gastric, Kidney, Leukemia, Melanoma, Ovarian, Osteosarcoma, Prostate, Soft Tissue Sarcoma, Thyroid, Urinary Bladder, Hepatobiliary, Contralateral cancer(s) not in pedigree, so they will be removed from the model specification.
#> Your model has 3 cancers - Endometrial, Pancreas, Small Intestine and 22 genes - ATM_hetero_anyPV, BARD1_hetero_anyPV, BRCA1_hetero_anyPV, BRCA2_hetero_anyPV, BRIP1_hetero_anyPV, CDH1_hetero_anyPV, CDK4_hetero_anyPV, CDKN2A[P16]_hetero_anyPV, CHEK2_hetero_1100delC, EPCAM_hetero_anyPV, MLH1_hetero_anyPV, MSH2_hetero_anyPV, MSH6_hetero_anyPV, MUTYH_hetero_anyPV, MUTYH_homo_anyPV, NBN_hetero_657del5, PALB2_hetero_anyPV, PMS2_hetero_anyPV, PTEN_hetero_anyPV, RAD51C_hetero_anyPV, RAD51D_hetero_anyPV, STK11_hetero_anyPV, TP53_hetero_anyPV.
#> ID 12's lower and upper bounds for the probability of carrying any pathogenic variant are (0.17495, 0.25534).
#> ID 19's lower and upper bounds for the probability of carrying any pathogenic variant are (0.10484, 0.14514).
#> ID 21's lower and upper bounds for the probability of carrying any pathogenic variant are (0.04897, 0.05066).
#> The value ranges in the results are driven by the different imputed missing ages at each imputation, so please consider including more age information when possible to reduce the range widths.
# Visualise the risk profile of the proband(s)
visRisk(output)