PanelPRO-11

Runs PanelPRO with 11 genes (genes = c("BRCA1", "BRCA2", "ATM", "PALB2", "CHEK2", "EPCAM", "PMS2", "MLH1", "MSH2", "MSH6", "CDKN2A")) and 11 cancers (cancers = c("Brain", "Breast", "Colorectal", "Contralateral", "Endometrial", "Gastric", "Kidney", "Melanoma", "Ovarian", "Pancreas", "Prostate", "Small intestine")), as in PanelPRO:::MODELPARAMS$PanPRO11.

PanelPRO11(...)

Arguments

...

Additional arguments to be passed into PanelPRO.

See also

Examples

data(test_fam_3)
PanelPRO11(test_fam_3, parallel = FALSE)
#> Contralateral cancer(s) not in pedigree, so they will be removed from the model specification.
#> Your model has 11 cancers - Brain, Breast, Colorectal, Endometrial, Gastric, Kidney, Melanoma, Ovarian, Pancreas, Prostate, Small Intestine and 11 genes - BRCA1_hetero_anyPV, BRCA2_hetero_anyPV, ATM_hetero_anyPV, PALB2_hetero_anyPV, CHEK2_hetero_1100delC, EPCAM_hetero_anyPV, PMS2_hetero_anyPV, MLH1_hetero_anyPV, MSH2_hetero_anyPV, MSH6_hetero_anyPV, CDKN2A[P16]_hetero_anyPV.
#> Setting max.mut to 1.
#> Riskmods that are NA will be removed
#> Warning: ID 14 is male but has been affected by OC; turning into non-affection.
#> ID 12's lower and upper bounds for the probability of carrying any pathogenic variant are (0.47446, 0.67114).
#> The value ranges in the results are driven by the different imputed missing ages at each imputation, so please consider including more age information when possible to reduce the range widths.
#> $posterior.prob
#> $posterior.prob$`12`
#>                       genes     estimate        lower        upper
#> 1                noncarrier 5.079653e-01 3.288555e-01 5.255387e-01
#> 2        BRCA1_hetero_anyPV 4.379404e-01 1.493927e-01 4.729691e-01
#> 3        BRCA2_hetero_anyPV 2.089101e-02 4.395298e-03 6.033246e-02
#> 4          ATM_hetero_anyPV 1.539978e-04 9.927370e-05 8.333875e-04
#> 5        PALB2_hetero_anyPV 8.222809e-05 3.562268e-05 7.036106e-04
#> 6     CHEK2_hetero_1100delC 1.824539e-04 1.190638e-04 9.991322e-04
#> 7        EPCAM_hetero_anyPV 5.594887e-08 4.435305e-08 6.182334e-08
#> 8         PMS2_hetero_anyPV 2.269498e-02 4.091221e-04 3.695656e-01
#> 9         MLH1_hetero_anyPV 1.977451e-03 6.142975e-05 3.286651e-02
#> 10        MSH2_hetero_anyPV 7.508253e-03 6.866653e-05 8.743094e-02
#> 11        MSH6_hetero_anyPV 5.999027e-04 2.563649e-05 7.761094e-03
#> 12 CDKN2A[P16]_hetero_anyPV 4.012857e-06 3.787432e-06 5.195862e-06
#> 
#> 
#> $future.risk
#> $future.risk$`12`
#> $future.risk$`12`$Brain
#>    ByAge     estimate        lower        upper
#> 1     49 0.0002952973 0.0002937912 0.0003192516
#> 2     54 0.0007028667 0.0006988741 0.0007577564
#> 3     59 0.0012485856 0.0012394334 0.0013490199
#> 4     64 0.0019108998 0.0018929167 0.0020763516
#> 5     69 0.0026965899 0.0026658514 0.0029591390
#> 6     74 0.0035699712 0.0034998103 0.0042557729
#> 7     79 0.0044174932 0.0043013343 0.0056032309
#> 8     84 0.0050694267 0.0049533093 0.0062549254
#> 9     89 0.0054638397 0.0053498999 0.0066255173
#> 10    94 0.0056133045 0.0055003567 0.0067638532
#> 
#> $future.risk$`12`$Breast
#>    ByAge     estimate        lower        upper
#> 1     49 0.0001340552 5.952009e-05 0.0003135558
#> 2     54 0.0003377318 1.559618e-04 0.0007756285
#> 3     59 0.0006745094 3.093538e-04 0.0015542413
#> 4     64 0.0011632840 5.279480e-04 0.0026939979
#> 5     69 0.0017577457 8.003927e-04 0.0040646596
#> 6     74 0.0023926210 1.098269e-03 0.0055123955
#> 7     79 0.0030171995 1.400437e-03 0.0069155768
#> 8     84 0.0035596854 1.671625e-03 0.0081146434
#> 9     89 0.0039037840 1.836526e-03 0.0088932621
#> 10    94 0.0040311183 1.898720e-03 0.0091793925
#> 
#> $future.risk$`12`$Colorectal
#>    ByAge    estimate       lower       upper
#> 1     49 0.002491268 0.001770059 0.009206154
#> 2     54 0.006369891 0.004908785 0.019953004
#> 3     59 0.010859462 0.008640843 0.031238466
#> 4     64 0.016187232 0.013234735 0.042903962
#> 5     69 0.022376499 0.018765336 0.054585747
#> 6     74 0.029044361 0.024893664 0.065631071
#> 7     79 0.035885294 0.031347703 0.075636065
#> 8     84 0.041945936 0.037192538 0.083604391
#> 9     89 0.046092610 0.041272147 0.088556087
#> 10    94 0.047910735 0.043099316 0.090480639
#> 
#> $future.risk$`12`$Endometrial
#>    ByAge estimate lower upper
#> 1     49        0     0     0
#> 2     54        0     0     0
#> 3     59        0     0     0
#> 4     64        0     0     0
#> 5     69        0     0     0
#> 6     74        0     0     0
#> 7     79        0     0     0
#> 8     84        0     0     0
#> 9     89        0     0     0
#> 10    94        0     0     0
#> 
#> $future.risk$`12`$Gastric
#>    ByAge     estimate        lower        upper
#> 1     49 0.0003010060 0.0002822977 0.0006004571
#> 2     54 0.0008132033 0.0007640118 0.0016005055
#> 3     59 0.0015814780 0.0014895834 0.0030520384
#> 4     64 0.0026194685 0.0024751699 0.0049247040
#> 5     69 0.0039746980 0.0037703970 0.0072272549
#> 6     74 0.0056311259 0.0053654777 0.0098370607
#> 7     79 0.0074254559 0.0071083569 0.0124180765
#> 8     84 0.0090367180 0.0086884557 0.0145061618
#> 9     89 0.0101613841 0.0098016586 0.0158101047
#> 10    94 0.0106809142 0.0103196243 0.0163535438
#> 
#> $future.risk$`12`$Kidney
#>    ByAge    estimate       lower       upper
#> 1     49 0.001018224 0.001016840 0.001032955
#> 2     54 0.002541791 0.002538364 0.002578237
#> 3     59 0.004592720 0.004586604 0.004657714
#> 4     64 0.007222096 0.007212646 0.007322387
#> 5     69 0.010367293 0.010354047 0.010507664
#> 6     74 0.013651921 0.013634916 0.013831775
#> 7     79 0.016623585 0.016603355 0.016837072
#> 8     84 0.018742807 0.018720361 0.018979156
#> 9     89 0.019993246 0.019969595 0.020241867
#> 10    94 0.020495222 0.020471355 0.020745964
#> 
#> $future.risk$`12`$Melanoma
#>    ByAge    estimate       lower       upper
#> 1     49 0.001531503 0.001529749 0.001549374
#> 2     54 0.003871594 0.003867277 0.003915597
#> 3     59 0.007325023 0.007317149 0.007405283
#> 4     64 0.012113650 0.012101213 0.012240425
#> 5     69 0.018539352 0.018521466 0.018721663
#> 6     74 0.025926913 0.025903804 0.026162460
#> 7     79 0.033310214 0.033283342 0.033584122
#> 8     84 0.039644424 0.039616263 0.039931476
#> 9     89 0.043788132 0.043762102 0.044053449
#> 10    94 0.045511758 0.045491060 0.045722734
#> 
#> $future.risk$`12`$Ovarian
#>    ByAge estimate lower upper
#> 1     49        0     0     0
#> 2     54        0     0     0
#> 3     59        0     0     0
#> 4     64        0     0     0
#> 5     69        0     0     0
#> 6     74        0     0     0
#> 7     79        0     0     0
#> 8     84        0     0     0
#> 9     89        0     0     0
#> 10    94        0     0     0
#> 
#> $future.risk$`12`$Pancreas
#>    ByAge     estimate        lower       upper
#> 1     49 0.0007190925 0.0006801005 0.001118765
#> 2     54 0.0021301666 0.0020235683 0.003261380
#> 3     59 0.0045108742 0.0042980112 0.006802888
#> 4     64 0.0080388388 0.0076773131 0.011939786
#> 5     69 0.0125292224 0.0119942426 0.018289379
#> 6     74 0.0174626771 0.0167622820 0.024999453
#> 7     79 0.0221367545 0.0213093667 0.031054508
#> 8     84 0.0257440949 0.0248393760 0.035482617
#> 9     89 0.0277032922 0.0267673345 0.037727459
#> 10    94 0.0283662753 0.0274224039 0.038425322
#> 
#> $future.risk$`12`$Prostate
#>    ByAge    estimate       lower       upper
#> 1     49 0.001812474 0.001802722 0.001859421
#> 2     54 0.007350717 0.007286433 0.007597945
#> 3     59 0.018551638 0.018369225 0.019210116
#> 4     64 0.036170843 0.035811755 0.037414485
#> 5     69 0.059248991 0.058687247 0.061101473
#> 6     74 0.080930337 0.080180917 0.083231281
#> 7     79 0.096991916 0.096167461 0.099362931
#> 8     84 0.106505728 0.105697804 0.108684234
#> 9     89 0.111552941 0.110795607 0.113437211
#> 10    94 0.113449079 0.112725004 0.115074078
#> 
#> $future.risk$`12`$`Small Intestine`
#>    ByAge     estimate        lower       upper
#> 1     49 0.0002004782 0.0001095000 0.001034267
#> 2     54 0.0004647077 0.0002710838 0.002228824
#> 3     59 0.0008006556 0.0004938056 0.003579862
#> 4     64 0.0012153537 0.0007986693 0.004968077
#> 5     69 0.0017116602 0.0011954047 0.006336173
#> 6     74 0.0022249139 0.0016290391 0.007535474
#> 7     79 0.0027319865 0.0020807111 0.008507680
#> 8     84 0.0031253568 0.0024462534 0.009127286
#> 9     89 0.0033659618 0.0026782559 0.009432806
#> 10    94 0.0034594735 0.0027717193 0.009523655
#> 
#> 
#>