PanelPRO-22

Runs PanelPRO with 22 genes (genes = c("ATM", "BARD1", "BRCA1", "BRCA2", "BRIP1", "CDH1", "CDK4", "CDKN2A", "CHEK2", "EPCAM", "MLH1", "MSH2", "MSH6", "MUTYH", "NBN", "PALB2", "PMS2", "PTEN", "RAD51C", "RAD51D", "STK11", "TP53")) and 17 cancers (cancers = c("Brain", "Breast", "Colorectal", "Contralateral", "Endometrial", "Gastric", "Kidney", "Leukemia", "Melanoma", "Ovarian", "Osteosarcoma", "Pancreas", "Prostate", "Small Intestine", "Soft Tissue Sarcoma", "Thyroid", "Urinary Bladder", "Hepatobiliary")), as in PanelPRO:::MODELPARAMS$PanPRO22.

PanelPRO22(...)

Arguments

...

Additional arguments to be passed into PanelPRO.

See also

Examples

data(test_fam_2)
PanelPRO22(test_fam_2, parallel = FALSE, max.mut = 1)
#> Brain, Breast, Colorectal, Gastric, Kidney, Leukemia, Melanoma, Ovarian, Osteosarcoma, Prostate, Soft Tissue Sarcoma, Thyroid, Urinary Bladder, Hepatobiliary, Contralateral cancer(s) not in pedigree, so they will be removed from the model specification.
#> Your model has 3 cancers - Endometrial, Pancreas, Small Intestine and 22 genes - ATM_hetero_anyPV, BARD1_hetero_anyPV, BRCA1_hetero_anyPV, BRCA2_hetero_anyPV, BRIP1_hetero_anyPV, CDH1_hetero_anyPV, CDK4_hetero_anyPV, CDKN2A[P16]_hetero_anyPV, CHEK2_hetero_1100delC, EPCAM_hetero_anyPV, MLH1_hetero_anyPV, MSH2_hetero_anyPV, MSH6_hetero_anyPV, MUTYH_hetero_anyPV, MUTYH_homo_anyPV, NBN_hetero_657del5, PALB2_hetero_anyPV, PMS2_hetero_anyPV, PTEN_hetero_anyPV, RAD51C_hetero_anyPV, RAD51D_hetero_anyPV, STK11_hetero_anyPV, TP53_hetero_anyPV.
#> ID 12's lower and upper bounds for the probability of carrying any pathogenic variant are (0.17495, 0.25534).
#> ID 19's lower and upper bounds for the probability of carrying any pathogenic variant are (0.10484, 0.14514).
#> ID 21's lower and upper bounds for the probability of carrying any pathogenic variant are (0.04897, 0.05066).
#> The value ranges in the results are driven by the different imputed missing ages at each imputation, so please consider including more age information when possible to reduce the range widths.
#> $posterior.prob
#> $posterior.prob$`12`
#>                       genes     estimate        lower        upper
#> 1                noncarrier 7.817906e-01 7.446588e-01 8.250531e-01
#> 2          ATM_hetero_anyPV 2.214493e-03 1.989672e-03 2.449508e-03
#> 3        BARD1_hetero_anyPV 4.131841e-04 3.717567e-04 4.594968e-04
#> 4        BRCA1_hetero_anyPV 6.801667e-04 6.111576e-04 7.524202e-04
#> 5        BRCA2_hetero_anyPV 7.687554e-04 6.902514e-04 8.491018e-04
#> 6        BRIP1_hetero_anyPV 7.004782e-04 6.302456e-04 7.789929e-04
#> 7         CDH1_hetero_anyPV 1.925507e-04 1.732449e-04 2.141332e-04
#> 8         CDK4_hetero_anyPV 2.555806e-04 2.299552e-04 2.842280e-04
#> 9  CDKN2A[P16]_hetero_anyPV 1.276994e-04 1.136785e-04 1.399614e-04
#> 10    CHEK2_hetero_1100delC 3.155536e-03 2.839151e-03 3.509232e-03
#> 11       EPCAM_hetero_anyPV 1.290308e-05 1.160937e-05 1.434936e-05
#> 12        MLH1_hetero_anyPV 7.670428e-02 6.393788e-02 9.475076e-02
#> 13        MSH2_hetero_anyPV 7.311322e-02 6.054324e-02 9.263085e-02
#> 14        MSH6_hetero_anyPV 2.362724e-02 1.989398e-02 2.784833e-02
#> 15       MUTYH_hetero_anyPV 1.931467e-03 6.078761e-04 2.792446e-03
#> 16         MUTYH_homo_anyPV 1.276352e-07 2.402887e-08 1.951223e-07
#> 17       NBN_hetero_657del5 4.622247e-04 4.158803e-04 5.140342e-04
#> 18       PALB2_hetero_anyPV 6.481422e-04 5.819550e-04 7.158827e-04
#> 19        PMS2_hetero_anyPV 1.018755e-03 9.166112e-04 1.132945e-03
#> 20        PTEN_hetero_anyPV 3.092720e-02 3.769888e-03 8.545749e-02
#> 21      RAD51C_hetero_anyPV 4.744854e-04 4.269117e-04 5.276692e-04
#> 22      RAD51D_hetero_anyPV 2.450752e-04 2.205030e-04 2.725450e-04
#> 23       STK11_hetero_anyPV 5.250571e-06 4.724129e-06 5.839093e-06
#> 24        TP53_hetero_anyPV 5.305378e-04 4.773441e-04 5.900043e-04
#> 
#> $posterior.prob$`19`
#>                       genes     estimate        lower        upper
#> 1                noncarrier 8.735553e-01 8.548611e-01 8.951552e-01
#> 2          ATM_hetero_anyPV 4.468706e-03 4.197773e-03 4.707536e-03
#> 3        BARD1_hetero_anyPV 4.842042e-04 4.552404e-04 5.170099e-04
#> 4        BRCA1_hetero_anyPV 1.308487e-03 1.116783e-03 1.390726e-03
#> 5        BRCA2_hetero_anyPV 1.851954e-03 1.559614e-03 1.961985e-03
#> 6        BRIP1_hetero_anyPV 8.208798e-04 7.717769e-04 8.764958e-04
#> 7         CDH1_hetero_anyPV 2.256473e-04 2.121497e-04 2.409353e-04
#> 8         CDK4_hetero_anyPV 2.995111e-04 2.815951e-04 3.198035e-04
#> 9  CDKN2A[P16]_hetero_anyPV 1.376319e-03 1.099975e-03 2.547509e-03
#> 10    CHEK2_hetero_1100delC 3.697925e-03 3.476725e-03 3.948466e-03
#> 11       EPCAM_hetero_anyPV 1.512093e-05 1.421644e-05 1.614540e-05
#> 12        MLH1_hetero_anyPV 3.997538e-02 3.336660e-02 4.903609e-02
#> 13        MSH2_hetero_anyPV 3.792342e-02 3.144180e-02 4.782539e-02
#> 14        MSH6_hetero_anyPV 1.220884e-02 1.027113e-02 1.439755e-02
#> 15       MUTYH_hetero_anyPV 1.043133e-03 3.699050e-04 1.479687e-03
#> 16         MUTYH_homo_anyPV 6.680205e-08 2.106848e-08 9.644340e-08
#> 17       NBN_hetero_657del5 5.416741e-04 5.092726e-04 5.783735e-04
#> 18       PALB2_hetero_anyPV 1.561393e-03 1.314920e-03 1.654161e-03
#> 19        PMS2_hetero_anyPV 1.193864e-03 1.122450e-03 1.274750e-03
#> 20        PTEN_hetero_anyPV 1.597707e-02 2.045899e-03 4.395684e-02
#> 21      RAD51C_hetero_anyPV 5.560423e-04 5.227813e-04 5.937151e-04
#> 22      RAD51D_hetero_anyPV 2.871999e-04 2.700204e-04 3.066582e-04
#> 23       STK11_hetero_anyPV 6.153064e-06 5.785005e-06 6.569946e-06
#> 24        TP53_hetero_anyPV 6.217292e-04 5.845390e-04 6.638524e-04
#> 
#> $posterior.prob$`21`
#>                       genes     estimate        lower        upper
#> 1                noncarrier 9.504277e-01 9.493442e-01 9.510307e-01
#> 2          ATM_hetero_anyPV 1.014859e-02 9.981118e-03 1.036858e-02
#> 3        BARD1_hetero_anyPV 6.064050e-04 6.012236e-04 6.122725e-04
#> 4        BRCA1_hetero_anyPV 2.728486e-03 2.573823e-03 2.764396e-03
#> 5        BRCA2_hetero_anyPV 4.344278e-03 4.093801e-03 4.400499e-03
#> 6        BRIP1_hetero_anyPV 1.028049e-03 1.019265e-03 1.037996e-03
#> 7         CDH1_hetero_anyPV 2.825949e-04 2.801803e-04 2.853293e-04
#> 8         CDK4_hetero_anyPV 3.751000e-04 3.718950e-04 3.787294e-04
#> 9  CDKN2A[P16]_hetero_anyPV 8.812108e-03 8.517758e-03 1.000076e-02
#> 10    CHEK2_hetero_1100delC 4.631188e-03 4.591617e-03 4.675998e-03
#> 11       EPCAM_hetero_anyPV 1.893707e-05 1.877526e-05 1.912030e-05
#> 12        MLH1_hetero_anyPV 5.597628e-03 5.466643e-03 6.630892e-03
#> 13        MSH2_hetero_anyPV 2.682520e-03 2.650722e-03 2.726813e-03
#> 14        MSH6_hetero_anyPV 2.780582e-04 2.760014e-04 2.805735e-04
#> 15       MUTYH_hetero_anyPV 1.306933e-04 1.296098e-04 1.318949e-04
#> 16         MUTYH_homo_anyPV 4.504487e-09 4.421374e-09 4.597318e-09
#> 17       NBN_hetero_657del5 6.783788e-04 6.725825e-04 6.849427e-04
#> 18       PALB2_hetero_anyPV 3.662686e-03 3.451507e-03 3.710086e-03
#> 19        PMS2_hetero_anyPV 1.495165e-03 1.482390e-03 1.509632e-03
#> 20        PTEN_hetero_anyPV 2.290413e-04 2.272366e-04 2.313136e-04
#> 21      RAD51C_hetero_anyPV 6.963732e-04 6.904231e-04 7.031112e-04
#> 22      RAD51D_hetero_anyPV 3.596818e-04 3.566086e-04 3.631621e-04
#> 23       STK11_hetero_anyPV 7.705941e-06 7.640099e-06 7.780503e-06
#> 24        TP53_hetero_anyPV 7.786378e-04 7.719848e-04 7.861718e-04
#> 
#> 
#> $future.risk
#> $future.risk$`12`
#> $future.risk$`12`$Endometrial
#>    ByAge   estimate       lower      upper
#> 1     49 0.01152550 0.009075558 0.01373355
#> 2     54 0.03178711 0.025775920 0.03792164
#> 3     59 0.05777037 0.047873012 0.06882647
#> 4     64 0.07823703 0.066263759 0.09251313
#> 5     69 0.08899290 0.076759942 0.10389718
#> 6     74 0.09510613 0.083090132 0.10975642
#> 7     79 0.09928160 0.087464124 0.11368881
#> 8     84 0.10203133 0.090350542 0.11628199
#> 9     89 0.10352426 0.091924401 0.11768919
#> 10    94 0.10407874 0.092511315 0.11821137
#> 
#> $future.risk$`12`$Pancreas
#>    ByAge     estimate        lower        upper
#> 1     49 0.0004933811 0.0004546522 0.0005504043
#> 2     54 0.0013058787 0.0012131727 0.0014433258
#> 3     59 0.0026497536 0.0024678054 0.0029200149
#> 4     64 0.0047222069 0.0043971974 0.0052044057
#> 5     69 0.0075996659 0.0070847026 0.0083635654
#> 6     74 0.0109242268 0.0102230474 0.0119668024
#> 7     79 0.0143464059 0.0135177524 0.0155858291
#> 8     84 0.0172555093 0.0163713477 0.0185876177
#> 9     89 0.0194033727 0.0185142486 0.0207528980
#> 10    94 0.0204863978 0.0196106862 0.0218217536
#> 
#> $future.risk$`12`$`Small Intestine`
#>    ByAge    estimate       lower       upper
#> 1     49 0.001228862 0.001035555 0.001518421
#> 2     54 0.002752120 0.002325647 0.003379406
#> 3     59 0.004492748 0.003808355 0.005468081
#> 4     64 0.006392966 0.005438455 0.007721408
#> 5     69 0.008240684 0.007035668 0.009895864
#> 6     74 0.009889407 0.008470349 0.011831652
#> 7     79 0.011144458 0.009573256 0.013305918
#> 8     84 0.012042585 0.010373342 0.014351080
#> 9     89 0.012476068 0.010767792 0.014843945
#> 10    94 0.012581334 0.010869795 0.014955048
#> 
#> 
#> $future.risk$`19`
#> $future.risk$`19`$Endometrial
#>    ByAge     estimate        lower        upper
#> 1     26 0.0001553814 0.0001005427 0.0002689382
#> 2     31 0.0005375780 0.0003512387 0.0009103970
#> 3     36 0.0014395054 0.0009563489 0.0023448633
#> 4     41 0.0034206289 0.0023968755 0.0051070120
#> 5     46 0.0075926816 0.0056940819 0.0100732521
#> 6     51 0.0156900268 0.0124482972 0.0189852357
#> 7     56 0.0286336941 0.0235635880 0.0331792854
#> 8     61 0.0430147991 0.0363665910 0.0497548639
#> 9     66 0.0534869842 0.0462735715 0.0612351617
#> 10    71 0.0599475330 0.0527783878 0.0677084072
#> 11    76 0.0645049630 0.0574558920 0.0721223145
#> 12    81 0.0677256407 0.0607667165 0.0752389747
#> 13    86 0.0697584489 0.0628583261 0.0772081669
#> 14    91 0.0707662146 0.0638969886 0.0781846388
#> 
#> $future.risk$`19`$Pancreas
#>    ByAge     estimate        lower        upper
#> 1     26 3.477325e-05 3.177288e-05 3.896332e-05
#> 2     31 9.409320e-05 8.640327e-05 1.049064e-04
#> 3     36 1.899916e-04 1.749387e-04 2.111678e-04
#> 4     41 3.547082e-04 3.274894e-04 3.930083e-04
#> 5     46 6.427314e-04 5.973089e-04 7.069638e-04
#> 6     51 1.139172e-03 1.069303e-03 1.238772e-03
#> 7     56 1.975596e-03 1.869515e-03 2.127642e-03
#> 8     61 3.300179e-03 3.135391e-03 3.536516e-03
#> 9     66 5.263932e-03 5.011451e-03 5.625637e-03
#> 10    71 7.839605e-03 7.482859e-03 8.350581e-03
#> 11    76 1.081230e-02 1.036605e-02 1.145271e-02
#> 12    81 1.381974e-02 1.332078e-02 1.453841e-02
#> 13    86 1.637365e-02 1.585674e-02 1.712130e-02
#> 14    91 1.810692e-02 1.759258e-02 1.885362e-02
#> 
#> $future.risk$`19`$`Small Intestine`
#>    ByAge     estimate        lower        upper
#> 1     26 9.395424e-05 7.940655e-05 0.0001143924
#> 2     31 2.842191e-04 2.395894e-04 0.0003479689
#> 3     36 6.058285e-04 5.105643e-04 0.0007434391
#> 4     41 1.079695e-03 9.116503e-04 0.0013238858
#> 5     46 1.672364e-03 1.415701e-03 0.0020463930
#> 6     51 2.390230e-03 2.029913e-03 0.0029155895
#> 7     56 3.273792e-03 2.791207e-03 0.0039661853
#> 8     61 4.281101e-03 3.665973e-03 0.0051444950
#> 9     66 5.363952e-03 4.616721e-03 0.0063943262
#> 10    71 6.420079e-03 5.553720e-03 0.0076023643
#> 11    76 7.339734e-03 6.378499e-03 0.0086484410
#> 12    81 8.063430e-03 7.037252e-03 0.0094657159
#> 13    86 8.556278e-03 7.493352e-03 0.0100123483
#> 14    91 8.765956e-03 7.692777e-03 0.0102373732
#> 
#> 
#> $future.risk$`21`
#> $future.risk$`21`$Endometrial
#>    ByAge     estimate        lower        upper
#> 1     24 0.0000325765 3.256406e-05 3.258948e-05
#> 2     29 0.0001382626 1.382149e-04 1.384075e-04
#> 3     34 0.0004095130 4.092940e-04 4.107715e-04
#> 4     39 0.0009412043 9.402225e-04 9.484022e-04
#> 5     44 0.0019049414 1.900886e-03 1.933638e-03
#> 6     49 0.0036061859 3.594024e-03 3.690353e-03
#> 7     54 0.0066454748 6.617960e-03 6.834287e-03
#> 8     59 0.0111390071 1.109190e-02 1.146127e-02
#> 9     64 0.0163087644 1.624808e-02 1.672389e-02
#> 10    69 0.0214295798 2.136535e-02 2.186971e-02
#> 11    74 0.0258143635 2.575032e-02 2.625413e-02
#> 12    79 0.0291360296 2.907240e-02 2.957356e-02
#> 13    84 0.0314060894 3.134275e-02 3.184207e-02
#> 14    89 0.0327056946 3.264252e-02 3.314069e-02
#> 15    94 0.0332434291 3.318034e-02 3.367793e-02
#> 
#> $future.risk$`21`$Pancreas
#>    ByAge     estimate        lower        upper
#> 1     24 1.545643e-05 1.540620e-05 1.580290e-05
#> 2     29 5.204770e-05 5.163437e-05 5.314264e-05
#> 3     34 1.153879e-04 1.143014e-04 1.186851e-04
#> 4     39 2.101149e-04 2.083689e-04 2.152204e-04
#> 5     44 4.095032e-04 4.057634e-04 4.213932e-04
#> 6     49 8.243709e-04 8.159037e-04 8.538230e-04
#> 7     54 1.532429e-03 1.517344e-03 1.586635e-03
#> 8     59 2.592127e-03 2.569915e-03 2.671842e-03
#> 9     64 4.112422e-03 4.081986e-03 4.219960e-03
#> 10    69 6.244814e-03 6.203939e-03 6.387394e-03
#> 11    74 8.853387e-03 8.802196e-03 9.031051e-03
#> 12    79 1.181336e-02 1.175482e-02 1.201698e-02
#> 13    84 1.454521e-02 1.448319e-02 1.476206e-02
#> 14    89 1.668679e-02 1.662410e-02 1.690696e-02
#> 15    94 1.781784e-02 1.775554e-02 1.803709e-02
#> 
#> $future.risk$`21`$`Small Intestine`
#>    ByAge     estimate        lower        upper
#> 1     24 1.281381e-05 1.262472e-05 1.428058e-05
#> 2     29 3.783933e-05 3.730417e-05 4.181769e-05
#> 3     34 7.825850e-05 7.720867e-05 8.572704e-05
#> 4     39 1.484002e-04 1.466187e-04 1.606501e-04
#> 5     44 2.533352e-04 2.506192e-04 2.716396e-04
#> 6     49 3.940429e-04 3.902391e-04 4.194337e-04
#> 7     54 5.911442e-04 5.860086e-04 6.253626e-04
#> 8     59 8.341138e-04 8.275039e-04 8.783148e-04
#> 9     64 1.151704e-03 1.143531e-03 1.206629e-03
#> 10    69 1.530070e-03 1.520416e-03 1.595241e-03
#> 11    74 1.932050e-03 1.921085e-03 2.006338e-03
#> 12    79 2.317749e-03 2.305792e-03 2.398951e-03
#> 13    84 2.660013e-03 2.647382e-03 2.745939e-03
#> 14    89 2.870173e-03 2.857254e-03 2.958145e-03
#> 15    94 2.954719e-03 2.941762e-03 3.042969e-03
#> 
#> 
#>