Calculate and format carrier probabilities and future risks

Estimates the posterior carrier probabilities using the peeling-paring algorithm and calculates future risk predictions for the proband(s). If missing ages were imputed, the average and range of the results from these imputations are reported.

PanelPROCalc(
  ped,
  lm,
  db,
  proband,
  ifmod = FALSE,
  net = FALSE,
  age.by = 5,
  max.mut = 2,
  debug = FALSE,
  rr.bcrat = NULL,
  rr.pop = rr.ref
)

Arguments

ped

A checked pedigree data frame returned by checkFam.

lm

A list of imputed ages returned by checkFam.

db

A model-specific database returned by buildDatabase.

proband

A numeric value or vector of the unique IDs in ped corresponding to the proband(s).

ifmod

A logical value indicating if cancer penetrance risk modifiers should be considered. The default is FALSE.

net

A logical value indicating whether to return net or crude future risk estimates. The default is FALSE, in which case crude future risk estimates will be returned.

age.by

The fixed age interval for the future risk calculation. The default is 5, in which case if the proband's current age is 30, risk predictions will be calculated at ages 35, 40, 45, etc.

max.mut

The maximum number of simultaneous mutations allowed. The default is 2.

debug

A logical value indicating whether to debug using browsers and sequential imputation. The default is FALSE.

rr.bcrat

A data frame with 3 columns: the IDs of the probands (ID), the probands' BCRAT relative risks for age < 50 (rr1), and the probands' BCRAT relative risks for age >= 50 (rr2). When net = FALSE and breast cancer is included in the model, these values will modify the future risk calculation. The default is NULL, in which case the BCRAT relative risks will not be incorporated.

rr.pop

A data frame of race-specific for estimates of 1 / (1 - population attributable risk). The default is rr.ref, which was estimated from NHIS 2015.

Value

A list with two components:

posterior.prob: A list where each element corresponds to a proband in proband and consists of a data frame of posterior carrier probabilities. There are three columns: estimate (the estimate if no ages were imputed, or the estimate averaged over all age imputations), lower (NA if no ages were imputed, or the minimum result for all age imputations), and upper (NA if no ages were imputed, or the maximum result for all age imputations). The rows correspond to the different genotypes. If no genes were in the model specification, each proband's data frame of carrier probabilities is replaced by the message "No carrier probabilities were requested by the model specification." If ages were imputed, ImputeRange messages giving the lower and upper bounds for the probability of carrying any pathogenic variant will be printed for each proband. A ZeroProb warning is raised if one or more of the estimated probabilities is 0 (or 1), germline testing results were incorporated, and default sensitivities/specifities of 1 were used.

future.risk: A list where each element corresponds to a proband in proband and consists of a list of data frames of future risk estimates at fixed intervals for each cancer in the model. Each data frame has three columns: estimate (the estimate if no ages were imputed, or the estimate averaged over all age imputations), lower (NA if no ages were imputed, or the minimum result for all age imputations), and upper (NA if no ages were imputed, or the maximum result for all age imputations). Each row corresponds to the risk estimate for at a different age, e.g. 35, 40, 45, etc. when age.by = 5 for a proband whose current age is 30. If no cancers were in the model specification, each proband's list of future risk estimates is replaced by the message "No future risk estimates were requested by the model specification."