Built-in PanelPRO database object

Built-in parameters for PanelPRO models. This object stores the default values, but the user can set their own parameter values or create their own database object with the same format.

PanelPRODatabase

Format

An object of class list of length 7.

Details

A list of 7 elements:

Penetrance: An array that provides the age-specific probabilities of developing a certain cancer, given genotype, race, and sex. The array has 6 levels corresponding to the cancer, gene, race, sex, age, and type of penetrance (net or crude). See str(PanelPRODatabase$Penetrance) for the structure of the array. For example, PanelPRODatabase$Penetrance["Breast", "BRCA1_hetero_anyPV", "Asian", "Female", "40", "Net"] is the net probability that an Asian woman with any heterozygous pathogenic BRCA1 variant develops breast cancer at age 40. For now, mutations in all genes are considered as heterozygous mutations except for mutations in MUTYH, which are allowed to be either heterozygous or homozygous.

AlleleFrequency: An array that provides the minor allele frequencies for mutations in each gene variant and ancestry type (Ashkenazi Jewish, non-Ashkenazi Jewish, and Italian). The array has 2 levels corresponding to gene and ancestry. See str(PanelPRODatabase$AlleleFrequency) for the structure of the array. For example, PanelPRODatabase$AlleleFrequency["BRCA1_anyPV", "AJ"] is the allele frequency of any pathogenic BRCA1 variant for an Ashkenazi Jewish individual.

DOC: An array that provides the age-specific probabilities of death by other causes (other than the specified cancer), given genotype, race, and sex. The array has 5 levels corresponding to the cancer, gene, race, sex, and age. See str(PanelPRODatabase$DOC) for the structure of the array. For example, PanelPRODatabase$DOC["Breast", "BRCA1_hetero_anyPV", "Asian", "Female", "40"] is the probability that a 40-year-old Asian women with any heterozygous pathogenic BRCA1 variant who is currently alive dies from a cause other than breast cancer. This is used for estimating crude future risk.

Riskmod: An array of parameters (either relative risks or hazard ratios) that modify risk at the cancer penetrance level. Risk modifiers are assumed to be independent of each other. The array has 6 levels corresponding to the cancer, gene, type of intervention (mastectomy, hysterectomy, oophorectomy), sex, age of intervention, and type of parameter (relative risk ratio or hazard ratio). See str(PanelPRODatabase$Riskmod) for the structure of the array. For example, PanelPRODatabase$Riskmod["Breast", "BRCA1_hetero_anyPV", "Mastectomy", "Female", "40", "HR"] is the ratio of the hazard of breast cancer for a woman with any heterozygous pathogenic BRCA1 variant who had a mastectomy at age 40 compared to the hazard for such a woman who did not have a mastectomy at age 40.

GermlineTesting: An array that provides the sensitivities and specificities for germline testing, used to modify the likelihood. The array has 2 levels corresponding to the gene and measure (sensitivity or specificity). See str(PanelPRODatabase$GermlineTesting) for the structure of the array. For example, PanelPRODatabase$GermlineTesting["BRCA1_hetero_anyPV", "sensitivity"] is the sensitivitiy for detecting BRCA1.

BiomarkerTesting: A list of parameters for tumor biomarker testing, where each component is an array corresponding to a different cancer. Each array represents a configuration of the tumor marker testing results (either "NoTest", "Neg", or "Pos") and the genotypes (0 or 1 for each gene that is related to the cancer). The parameters represent the probability of the tumor marker configuration given genotype. See str(PanelPRODatabase$BiomarkerTesting) for the structure of the list and arrays. For example, PanelPRODatabase$BiomarkerTesting$Breast[ER="Pos", CK5.6="Pos", CK14="Pos", PR="Neg", HER2="Neg", BRCA1_hetero_anyPV="1", BRCA2_hetero_anyPV="0"] is the probability of being a BRCA1 pathogenic variant carrier but not a BRCA2 pathogenic variant carrier given that the individual tested positive for ER, CK5.6, and CK14 and negative for PR and HER2.

Contralateral: An array that provides the age-specific probabilities of developing contralateral breast cancer for carriers of pathogenic germline mutations of CBC associated genes and who have already had a first primary breast cancer in one breast, given genotype, sex, and age of the first breast cancer diagnosis. The array has 5 levels corresponding to the gene, sex, age at first breast cancer, type of penetrance (net or crude), and age. See str(PanelPRODatabase$Contralateral) for the structure of the array. For example, PanelPRODatabase$Penetrance["BRCA1_hetero_anyPV", "Female", "40", "Net", "50"] is the net probability that a woman with any heterozygous pathogenic BRCA1 variant, who had their first breast cancer at age 40, develops contralateral breast cancer at age 50. For now, mutations in the CBC associated genes are considered as heterozygous mutations.

References

Antoniou, A. C., et al. "A comprehensive model for familial breast cancer incorporating BRCA1, BRCA2 and other genes." British journal of cancer 86.1 (2002): 76-83.

Begg, Colin B., et al. "Lifetime risk of melanoma in CDKN2A mutation carriers in a population-based sample." Journal of the National Cancer Institute 97.20 (2005): 1507-1515.

Berwick, Marianne, et al. "The prevalence of CDKN2A germ-line mutations and relative risk for cutaneous malignant melanoma: an international population-based study." Cancer Epidemiology and Prevention Biomarkers 15.8 (2006): 1520-1525.

Bishop, D. Timothy, et al. "Geographical variation in the penetrance of CDKN2A mutations for melanoma." Journal of the National Cancer Institute 94.12 (2002): 894-903.

Braun, Danielle, et al. "A clinical decision support tool to predict cancer risk for commonly tested cancer-related germline mutations." Journal of genetic counseling 27.5 (2018): 1187-1199.

Chen, Jinbo, et al. "Penetrance of Breast and Ovarian Cancer in Women Who Carry a BRCA1/2 Mutation and Do Not Use Risk-Reducing Salpingo-Oophorectomy: An Updated Meta-Analysis." JNCI cancer spectrum 4.4 (2020): pkaa029.

Chen, Sining, et al. "BayesMendel: an R environment for Mendelian risk prediction." Statistical applications in genetics and molecular biology 3.1 (2004).

Chen, Sining, et al. "Prediction of germline mutations and cancer risk in the Lynch syndrome." Jama 296.12 (2006): 1479-1487.

Chowdhury, M., et al. “A Model for Individualized Risk Prediction of Contralateral Breast Cancer”, Breast Cancer Research and Treatment 161 (2017): 153-160.

Chu, Haitao, Sining Chen, and Thomas A. Louis. "Random effects models in a meta-analysis of the accuracy of two diagnostic tests without a gold standard." (2007).

DevCan: Probability of Developing or Dying of Cancer Software, Version 6.7 Surveillance Research Program, Statistical Methodology and Applications, National Cancer Institute, 2012. http://surveillance.cancer.gov/devcan/

Felton, K. E. A., D. M. Gilchrist, and S. E. Andrew. "Constitutive deficiency in DNA mismatch repair: is it time for Lynch III?." Clinical genetics 71.6 (2007): 499-500.

Guo, Yonghai, et al. "Risk of ipsilateral breast tumor recurrence and contralateral breast cancer in patients with and without TP53 variant in a large series of breast cancer patients." The Breast 65 (2022): 55-60.

Katki, Hormuzd A. "Incorporating medical interventions into carrier probability estimation for genetic counseling." BMC medical genetics 8.1 (2007): 1-11.

Kuchenbaecker, Karoline B., et al. "Risks of breast, ovarian, and contralateral breast cancer for BRCA1 and BRCA2 mutation carriers." Jama 317.23 (2017): 2402-2416.

Lakhani, Sunil R., et al. "The pathology of familial breast cancer: predictive value of immunohistochemical markers estrogen receptor, progesterone receptor, HER-2, and p53 in patients with mutations in BRCA1 and BRCA2." Journal of clinical oncology 20.9 (2002): 2310-2318.

Lakhani, Sunil R., et al. "Prediction of BRCA1 status in patients with breast cancer using estrogen receptor and basal phenotype." Clinical cancer research 11.14 (2005): 5175-5180.

Lee, Andrew J., et al. "Incorporating truncating variants in PALB2, CHEK2, and ATM into the BOADICEA breast cancer risk model." Genetics in Medicine 18.12 (2016): 1190-1198.

Rosenthal, Eric T., et al. "Clinical testing with a panel of 25 genes associated with increased cancer risk results in a significant increase in clinically significant findings across a broad range of cancer histories." Cancer genetics 218 (2017): 58-68.

Sajal, I.H., et al. "CBCRisk-Black: A Personalized Contralateral Breast Cancer Risk Prediction Model for Black Women", Breast Cancer Research and Treatment 194 (2022): 179-186.

Theodoratou, Evropi, et al. "A large-scale meta-analysis to refine colorectal cancer risk estimates associated with MUTYH variants." British journal of cancer 103.12 (2010): 1875-1884.

Wang, Cathy, et al. "Penetrance of colorectal cancer among mismatch repair gene mutation carriers: a meta-analysis." JNCI cancer spectrum 4.5 (2020): pkaa027.

Wang, Wenyi, et al. "Estimating CDKN2A carrier probability and personalizing cancer risk assessments in hereditary melanoma using MelaPRO." Cancer research 70.2 (2010): 552-559.

Win, Aung Ko, et al. "Risk of extracolonic cancers for people with biallelic and monoallelic mutations in MUTYH." International journal of cancer 139.7 (2016): 1557-1563.

See also