Impute missing ages in pedigree

Imputes missing current, cancer affection, and death ages for relatives in ped, subject to age-bounding information.

ageImpute(
  ped,
  rel_l,
  proband,
  impute_times,
  max_iter_tries,
  penet_db,
  dist_age = AgeDistribution,
  penet_cbc_db = NULL
)

Arguments

ped

A pedigree data frame returned by .pedStructureCheck.

rel_l

The list of relationships in ped returned by .pedStructureCheck.

proband

A numeric value or vector of the unique IDs in ped corresponding to the proband(s).

impute_times

The number of iterations to be imputed for missing ages.

max_iter_tries

A numeric value indicating the maximum number of iterations that should be attempted when multiply imputing missing ages.

penet_db

The cancer penetrances for carriers returned by buildDatabase in the penet$penet_c list component.

dist_age

A list containing empirical distributions of female and male ages at first birth; female and male birth ages; and sibling age differences. The default is AgeDistribution.

penet_cbc_db

A numeric array of contralateral cancer penetrance for carriers returned by buildDatabase in the penet$penet_cbc list component.

Value

A list where each component is a matrix of imputed ages corresponding to CurAge (if one or more individuals in ped is missing their current age) or AgeX, where X is a cancer for which one or more individuals in ped is missing the affection age. Each matrix has impute_times rows; the column names are the IDs of the individuals for whom ages were imputed.

Details

First, we impute the missing censoring ages (stored as ped$CurAge) of living first-degree relatives, followed by second-degree relatives and then more distant relatives. If more than one proband is supplied, the first one will be used to determine first- and second-degree relatives.

Within each group of relatives, we estimate current age bounds for each individual (see .getAgeFromPed documentation). The relative with the narrowest age bound interval in the relative group has their current age imputed first, from a normal distribution truncated by their age bounds (see .currentAgeImpute documentation). The pedigree is updated with this relative's imputed age and the age bounds are then re-calculated for the remaining relatives in the relative group. This process repeats until no living individuals with missing censoring ages are left in the relative group.

After all missing censoring ages for living relatives have been imputed, we impute missing cancer affection ages. For each missing age, the upper bound is the minimum of the individual's censoring age, their upper bound for this cancer's affection age, and PanelPRO:::MAXAGE; the lower bound is the maximum of their lower bound for this cancer's affection age, their ages of preventative interventions (if relevant), and 1. The affection age is then sampled based on the crude SEER cancer penetrance estimates (subset by the individual's race and sex) within the age bounds (see .cancerAgeImpute documentation). If a missing cancer age cannot be sampled such that it satisfies its age bounds, this imputation run will be discarded and re-attempted, with a warning message issued. If the desired number of runs (iterations, default 20) cannot be imputed within max.iter.tries (default 5*iterations) attempts, an error will be raised.

Finally, missing death ages (stored in CurAge in the pedigree) of individuals who are affected for at least one cancer and/or reported at least one intervention are imputed, using the maximum of their cancer and intervention ages as the lower bound (see .currentAgeImpute documentation). If a lower imputation bound cannot be found using this approach and the individual has children in the pedigree, we assign their CurAge to be 15 plus the maximum difference in their living children's ages in the current imputation. (If the individual has one or zero living children, CurAge is simply set to 15.) In cases where the individual has no children in the pedigree, their CurAge is set to 1.

See also