Build model-specific database

Build a database of parameters for only the genes and cancers in the model. The genes and cancers are taken from the model specification model_spec or user-supplied vectors of genes and cancers. Allele frequencies, risk modifiers, and the net cancer/death by other causes penetrances for carriers are subsetted directly from the master database, ppd. The crude cancer penetrances for carriers are calculated from the net cancer penetrances for carriers and the penetrances for death by other causes. The noncarrier cancer and death by other causes penetrances are calculated from the carrier penetrances and the SEER penetrances.

buildDatabase(
  model_spec = NULL,
  ppd = PanelPRO::PanelPRODatabase,
  genes = NULL,
  cancers = NULL
)

Arguments

model_spec

One of the following character strings indicating a pre-specified model: "BRCAPRO", "BRCAPRO5", "BRCAPRO6", "MMRPRO", "PanPRO11", and "PanPRO22". See PanelPRO:::MODELPARAMS for more information. Either model_spec or genes and cancers should be specified by the user. When both input types are supplied, the input from genes and cancers will take precedence. The default is NULL, for no model specified.

ppd

A structured list providing the parameters of the model. The default is PanelPRODatabase. Users who wish to provide their own model parameters should supply a list with the same structure as PanelPRODatabase.

genes

A character vector of the genes of interest. The default is NULL. Available options are listed in PanelPRO:::GENE_TYPES. Either model_spec or genes and cancers should be specified by the user. When both input types are supplied, the input from genes and cancers will take precedence. It is considered valid to specify no genes in genes as long as at least one cancer is specified in cancers.

cancers

A character vector of the cancers of interest. Available options are listed in PanelPRO:::CANCER_TYPES. The default is NULL. Either model_spec or genes and cancers should be specified by the user. When both input types are supplied, the input from genes and cancers will take precedence. It is considered valid to specify no cancers in cancers as long as at least one gene is specified in genes.

Value

A list of 7 elements containing the model parameters:

af: Allele frequencies for the genes. Can be subsetted by ethnicity.

penet: Cancer penetrances, split into penet_c for mutation carriers, penet_nc for noncarriers and penet_cbc for contralateral breast cancer penetrances for carriers and noncarriers. penet_c and penet_nc can be subset by cancer, gene, race, sex, age, and type of penetrance (net or crude). penet_cbc is a list of two arrays, cbc_noncarrier for female non-carriers and cbc_carrier for carriers of any sex. The CBC non-carrier array can be subsetted by gene, age of first breast cancer, type of first breast cancer, if first degree relatives had breast cancer, ER result, age category at first child birth, if anti-estrogen therapy was used to treat the first breast cancer, history of high risk preneoplasia, breast density, net or crude penetrance, and age. The CBC carrier array can be subsetted by gene, sex, age of first breast cancer, net or crude penetrance, and age.

riskmod: Information for interventions, such as mastectomies, oophorectomies, and hysterectomies, that impact the cancer penetrances. Can be subsetted by cancer, gene, intervention, sex, age of intervention, and data type (relative risk or hazard ratio).

doc: Probability of dying of causes other than the cancer of interest, split into doc_c for mutation carriers and doc_nc for noncarriers. Each can be subsetted by cancer, gene, race, sex, and age.

germline: Sensitivity and specificity for germline testing, which can be used to modify the likelihood.

biomarker: Information on tumor biomarker testing, which can be used to modify the likelihood.

MS: List of genes and cancers specified by the model.

Examples

brcadb1 <- buildDatabase(model_spec = "BRCAPRO")
#> Your model has 3 cancers - Breast, Ovarian, Contralateral and 2 genes - BRCA1_hetero_anyPV, BRCA2_hetero_anyPV.
brcadb2 <- buildDatabase(genes = c("BRCA1", "BRCA2"), 
                         cancers = c("Breast", "Ovarian"))
#> Your model has 2 cancers - Breast, Ovarian and 2 genes - BRCA1_hetero_anyPV, BRCA2_hetero_anyPV.
stopifnot(all.equal(brcadb1, brcadb2))
#> Error in eval(expr, envir, enclos): brcadb1 and brcadb2 are not equal:
#>   Component "penet": Component "penet_cbc": Modes: numeric, NULL
#>   Component "penet": Component "penet_cbc": Lengths: 70688, 0
#>   Component "penet": Component "penet_cbc": Attributes: < Modes: list, NULL >
#>   ....