Calculate BCRAT relative risks

Calculate BCRAT relative risks

calc.rr.bcrat(df)

Arguments

df

A data frame of BCRAT covariates, where each row corresponds to a proband in the pedigree. The expected columns are:

ID: Numeric ID for each proband. Should not contain duplicated entries.

Age: Initial age (should be >=20 and <90).

NBiops: Number of benign breast biopsies (99 if unknown).

Hyp: A logical value indicating whether the individual has had a breast biopsy with atypical hyperplasia (0 if no, 1 if yes, 99 if unknown).

AgeMen: Age at menarche (99 if unknown)

AgeFLB: Age at first live birth (98 if nulliparous, 99 if unknown).

NumRel: The number of female first-degree relatives with breast cancer (99 if unknown).

Race: A character string indicating race ("White", "Black", "Hispanic", "FHispanic", "Asian", or "Other"; "Hispanic" corresponds to U.S.-born Hispanic while "FHispanic" corresponds to foreign-born Hispanic).

Value

A data frame of BCRAT relative risk estimates with three columns:

ID: Numeric ID for each individual.

rr1: Individual's BCRAT relative risk estimate for ages < 50.

rr2: Individual's BCRAT relative risk estimate for ages >= 50.

Details

The model for White women is described in Gail et al. (1989) and Costantino et al. (1999); the model for African-American women is described in Gail et al. (2007); the model for Asian women is described in Matsuno et al. (2011); and the model for Hispanic women is described in Banegas et al. (2017).

References

Gail MH, Brinton LA, Byar DP, Corle DK, Green SB, Schairer C, Mulvihill JJ. Projecting individualized probabilities of developing breast cancer for white females who are being examined annually. JNCI: Journal of the National Cancer Institute. 1989 Dec 20;81(24):1879-86.

Costantino JP, Gail MH, Pee D, Anderson S, Redmond CK, Benichou J, Wieand HS. Validation studies for models projecting the risk of invasive and total breast cancer incidence. Journal of the National Cancer Institute. 1999 Sep 15;91(18):1541-8.

Gail MH, Costantino JP, Pee D, Bondy M, Newman L, Selvan M, Anderson GL, Malone KE, Marchbanks PA, McCaskill-Stevens W, Norman SA. Projecting individualized absolute invasive breast cancer risk in African American women. Journal of the National Cancer Institute. 2007 Dec 5;99(23):1782-92.

Matsuno RK, Costantino JP, Ziegler RG, Anderson GL, Li H, Pee D, Gail MH. Projecting individualized absolute invasive breast cancer risk in Asian and Pacific Islander American women. Journal of the National Cancer Institute. 2011 Jun 22;103(12):951-61.

Banegas MP, John EM, Slattery ML, Gomez SL, Yu M, LaCroix AZ, Pee D, Chlebowski RT, Hines LM, Thompson CA, Gail MH. Projecting individualized absolute invasive breast cancer risk in US Hispanic women. Journal of the National Cancer Institute. 2017 Feb 1;109(2):djw215.

See also

Examples

bcrat.input <- data.frame(
  ID = 1:6, Age = 45:50, NBiops = c(0, 0, 0, 99, 1, 3),
  Hyp = c(99, 99, 99, 99, 0, 1), AgeMen = c(11, 12, 13, 14, 99, 15),
  AgeFLB = c(20, 22, 98, 99, 24, 30), NumRel = c(0, 1, 2, 0, 0, 0),
  Race = c("White", "White", "Black", "Hispanic", "Asian", "Other")
)
PanelPRO:::calc.rr.bcrat(bcrat.input)
#>   ID       rr1      rr2
#> 1  1  1.501774 1.501774
#> 2  2  2.945056 2.945056
#> 3  3  3.382792 3.382792
#> 4  4  1.000000 1.000000
#> 5  5  2.130695 2.130695
#> 6  6 10.107154 5.681178