Calculate an individual's cancer density probabilities

Calculates the cancer density functions for an individual.

calcCancerDens(
  ID,
  affages,
  cancers,
  sub_dens,
  PGs,
  direct_fill_PGs,
  multi_PGs,
  multi_muts
)

Arguments

ID

Unique ID corresponding to an ID in sub_dens.

affages

Individual ID's (non-missing) ages of affection corresponding to cancers.

cancers

A vector of cancer names for cancers that individual ID developed at affages.

sub_dens

A numeric array of subsetted cancer penetrances returned by subsetCancerPenetrance.

PGs

Possible genotypes in both list and data frame format, returned by .getPossibleGenotype.

direct_fill_PGs

A vector of genotype names with no more than 1 mutation (including "noncarrier"), corresponding to PGs.

multi_PGs

A vector of genotype names with 2 or more simultaneous mutations, corresponding to PGs.

multi_muts

A list where each component corresponds to a genotype in multi_PGs, represented as a vector of gene names to indicate the mutations. mutations, corresponding to PGs.

Value

A numeric vector with length equal to the number of possible genotypes. Each element corresponds to the probability that individual ID

develops the cancers at their corresponding affages, given a possible genotype in PGs. If cancers is an empty vector, 1 will be returned.