Calculate penetrances

Calculates the density and survival penetrance functions (for either cancer or for death by other causes) for each proband in the pedigree, based on the genes and cancers in the model.

calcCancerPenetrance(
  fam,
  proband,
  db = NULL,
  sub_dens = NULL,
  PGs,
  direct_fill_PGs,
  multi_PGs,
  multi_muts,
  consider.modification = FALSE,
  net = TRUE,
  doc = FALSE,
  lm.ages = NULL
)

Arguments

fam

A checked pedigree data frame returned by checkFam.

proband

A numeric value or vector of the unique IDs in ped for whom to return cancer penetrances.

db

A model-specific database returned by buildDatabase. The default is NULL, in which case the penetrances in sub_dens will be used directly instead of subsetted from db. One of db and sub_dens must be non-NULL.

sub_dens

A numeric array of subsetted cancer penetrances returned by subsetCancerPenetrance. The default is NULL, in which case the penetrances will be subsetted from db. One of db and sub_dens must be non-NULL.

PGs

Possible genotypes in both list and data frame format, returned by .getPossibleGenotype.

direct_fill_PGs

A vector of genotype names with no more than 1 mutation (including "noncarrier"), corresponding to PGs.

multi_PGs

A vector of genotype names with 2 or more simultaneous mutations, corresponding to PGs.

multi_muts

A list where each component corresponds to a genotype in multi_PGs, represented as a vector of gene names to indicate the mutations. mutations, corresponding to PGs.

consider.modification

A logical value indicating whether or not to consider interventions that modify the penetrance. The default is FALSE.

net

A logical value indicating if net or crude penetrances should be returned. The default is TRUE, for net penetrances.

doc

A logical value indicating whether the pentrances for death by other causes should be calculated. If FALSE, the cancer penetrances will be calculated instead. The default is FALSE.

lm.ages

A list of imputed ages returned by checkFam. Default is NULL; only needed when CBC is in the model and a BC age is missing.

Value

A list of density (Dens) and survival (Sur) penetrance arrays for the probands, each of which contains the following dimensions:

ID: The IDs of the probands in the family (from ID in fam).

cancers: The cancers included in the model.

genotype: All possible genotypes, with a maximum number of simultaneous mutations based on max_mut.

ages: Ages 1-PanelPRO:::MAXAGE.