Calculate future risk

Calculates the future risk of cancer for the proband(s), based on the posterior carrier probabilities.

calcFutureRisk(
  ped,
  proband,
  MS_cancers,
  cplist,
  doclist,
  posterior_probs,
  net = FALSE,
  age.by = 5,
  rr.bcrat = NULL,
  rr.pop = rr.ref,
  cbc_penets,
  lm.ages
)

Arguments

ped

A checked pedigree data frame returned by checkFam.

proband

A numeric value or vector of the unique IDs in ped for whom to return future risks.

MS_cancers

The cancers in the model, obtained from the model-specific database.

cplist

A list containing the cancer penetrances for the probands in ped, returned by calcCancerPenetrance. If net == TRUE, these should be net penetrances; if net == FALSE, these should be crude.

doclist

A list containing the penetrances for the death by other causes for ped, returned by calcCancerPenetrance.

posterior_probs

The genotypic distribution matrix of the family members in proband, returned by pp.peelingParing.

net

A logical value indicating whether to use the net cancer penetrances to estimate future risk. The default is FALSE, in which case the crude penetrances will be used.

age.by

The number of years between each future risk prediction. The default is 5; if the proband's current age is 30, risk predictions will be calculated at ages 35, 40, 45, etc.

rr.bcrat

A data frame with 3 columns: the IDs of the probands ("ID"), the probands' BCRAT relative risks for age < 50 ("rr1"), and the probands' BCRAT relative risks for age >= 50 ("rr2"). When net = FALSE and breast cancer is included in the model, these values will modify the future risk calculation. The default is NULL, in which the BCRAT relative risks will not be incorporated.

rr.pop

A data frame of race-specific for estimates of 1 / (1 - population attributable risk). The default is rr.ref, which was estimated from NHIS 2015.

cbc_penets

an array of CBC carrier penetrances as returned by the penet$penet_cbc component of buildDatabase.

lm.ages

imputed ages for the pedigree as returned by checkFam.

Value

A list of arrays, one for each proband, storing the future risks of developing the cancers specified by the model.