Check and make corrections to pedigree

Check and make corrections to pedigree

checkFam(
  ped,
  db = NULL,
  proband = NULL,
  unknown.race = UNKNOWN_RACE,
  unknown.ancestry = UNKNOWN_ANCESTRY,
  ignore.proband.germ = FALSE,
  allow.age.zero = FALSE,
  impute.missing.ages = TRUE,
  impute.times = 20,
  max.iter.tries = NULL,
  random.seed = 42L
)

Arguments

ped

A pedigree data frame. (See the description of the pedigree parameter in the PanelPRO master function documentation.)

db

A model-specific database returned by buildDatabase. The default is NULL, in which case the database will be constructed based on the cancer information encoded in ped and the genes in GENE_TYPES.

proband

A numeric value or vector of the unique IDs in ped corresponding to the proband(s). The default is NULL, in which case probands will be inferred from the isProband column in ped, if it exists and is valid. Otherwise, the IDs in proband will override the contents of ped$isProband. The user can also specify proband = "All", in which case all IDs in the ped will be treated as probands.

unknown.race

A character string indicating the default race to use when ped$race is missing or unsupported. The default is PanelPRO:::UNKNOWN_RACE.

unknown.ancestry

A character string indicating the default ancestry to use when ped$Ancestry is missing or unsupported. The default is PanelPRO:::UNKNOWN_ANCESTRY.

ignore.proband.germ

A logical value indicating if the proband(s)'s germline testing results (if provided in ped) should be considered. The default is FALSE, in which case the proband(s)'s germline testing results will not be ignored.

allow.age.zero

A logical value indicating whether zero ages can be reported in ped. The default is FALSE, in which case 0 values in age columns will be treated as missing and imputed. allow.age.zero should only be TRUE if individuals are confirmed to born with a cancer. Currently, the allow.age.zero = TRUE is not supported behavior for the PanelPRO model functions.

impute.missing.ages

A logical value indicating whether missing ages should be multiply imputed. The default is TRUE. If set to FALSE, an error will be raised if any individuals have missing current or cancer affection ages.

impute.times

A numeric value indicating the number of iterations that should be run when multiply imputing missing ages. The default is 20.

max.iter.tries

A numeric value indicating the maximum number of iterations that should be attempted when multiply imputing missing ages. (Invalid iterations typically occur when an individual is missing both their current and cancer ages. Current ages for living individuals are imputed first; if a valid cancer affection cannot be found, subject to the upper bound set by the imputed current age, this imputation run will be discarded.) The default is NULL, in which case the maximum number of tries will be set to five times iterations.

random.seed

The random seed (a numeric value) to set for imputing missing ages. The default is 42L.

Value

A list with three components:

lms: A nested list where each component corresponds to the imputed ages for a connected pedigree in ped_list. Each component is itself a list where each component is a matrix of imputed ages corresponding to CurAge (if one or more individuals in ped is missing their current age) or AgeX, where X is a cancer for which one or more individuals in ped is missing the affection age. Each matrix has impute_times rows; the column names are the IDs of the individuals for whom ages were imputed.

proband: A numeric value or vector of the unique IDs in the pedigree corresponding to the proband(s). If the proband argument was specified by the user, the returned list component will be identical. Otherwise, the IDs will correspond to individuals in the ped argument where ped$isProband == 1.

ped_list: A list where each element is a connected pedigree that is a subset of the relatives in ped. Typically, ped_list has length 1, unless there are multiple probands who are related to multiple disconnected pedigrees. Relatives who are not connected to the proband(s) will be removed.

Details

First, checkFam checks that all required columns in pedigree are present, valid, and consistent, printing informative error/warning messages and making corrections when possible. It also removes family members that are not connected to the proband(s) and splits the pedigree into connected sub-families, if necessary. For more details, see the .pedStructureCheck documentation.

Second, checkFam imputes missing current, cancer affection, and death ages for relatives in ped, subject to age-bounding information. For more details, see the ageImpute documentation.

See also

Examples

# Build database for specific model configurations
brcadb <- buildDatabase("BRCAPRO")
#> Your model has 3 cancers - Breast, Ovarian, Contralateral and 2 genes - BRCA1_hetero_anyPV, BRCA2_hetero_anyPV.
# Run pedigree checks and age imputations
checkFam(ped = test_fam_1, db = brcadb)
#> Error in .pedStructureCheck(ped, cancers = MS_cancers, proband = proband,     unknown.race = unknown.race, unknown.ancestry = unknown.ancestry,     ignore.proband.germ = ignore.proband.germ, allow.age.zero = allow.age.zero): CBC cancer is required in model but cannot be found in the pedigree