Impute current age for an individual

Impute current age for an individual

.currentAgeImpute(rel_ages, ab, dist_age = AgeDistribution)

Arguments

rel_ages

A list of the individual's current age-bounding information returned by .getAgeFromPed.

ab

A numeric vector of length two with the individual's current age bounds.

dist_age

A list containing empirical distributions of female and male ages at first birth; female and male birth ages; and sibling age differences. The default is AgeDistribution.

Value

An integer value representing the imputed current age.

Details

Samples current age based on the empirical age distributions in dist_age, subject to the age bounds in ab and age information in rel_ages.

First, attempt to impute the following ages when the age information is known:

If the individual's mother's current age is known, calculate the difference between the mother's age and an age randomly sampled from the distribution of female birth ages

If the individual's father's current age is known, calculate the difference between the father's age and an age randomly sampled from the distribution of male birth ages

If at least one of the individual's children's current ages is known, calculate the sum of the oldest child's current age and an age randomly sampled from the distribution of female ages at first birth, if the individual is female; calculate the sum of the oldest child's current age and an age randomly sampled from the distribution of male ages at first birth, if the individual is male

If at least one of the individual's siblings' current ages is known, calculate the sum of the mean sibling age and +/- an age difference randomly sampled from the distribution of sibling age differences, where the +/- sign is determined by a fair coin toss

The final imputed age returned by the function is the mean of the ages successfully imputed by the above approaches, bounded by the lower and upper age bounds in ab. If no ages were successfully imputed by any of the approaches, sample from a uniform distribution bounded by the lower and upper age bounds in ab.

See also