Calculate carrier probabilities and future risks for a single age imputation

This function is called within PanelPROCalc to calculate values when multiply imputing missing ages. It wraps the routine that can be executed in parallel, allowing for flexibility when toggling the parallel argument. It is also useful for debugging (debug = TRUE), in which the non-parallel version should be executed.

.parallelCalc(
  tt,
  lm,
  ped,
  db,
  sub_dens_lik,
  cplist.fr,
  PGs,
  direct_fill_PGs,
  multi_PGs,
  multi_muts,
  germ,
  marker,
  proband,
  max.mut,
  af,
  net,
  rr.bcrat,
  rr.pop,
  doclist,
  age.by,
  homo_idx
)

Arguments

tt

A numeric value that indexes which set of imputed ages in lm is being used in the calculation.

lm

A list of imputed ages returned by checkFam.

ped

A checked pedigree data frame returned by checkFam.

db

A model-specific database returned by buildDatabase.

sub_dens_lik

A numeric array of subsetted cancer penetrances returned by subsetCancerPenetrance, to be used for likelihood calculation.

cplist.fr

A list of density and survival arrays for cancer penetrances for the probands in proband, returned by calcCancerPenetrance. (If net == TRUE, they should be net penetrances, and if net == FALSE, they should be crude.) They will be used to calculate future risk.

PGs

Possible genotypes in both list and data frame format, returned by .getPossibleGenotype.

direct_fill_PGs

A vector of genotype names with no more than 1 mutation (including "noncarrier"), corresponding to PGs.

multi_PGs

A vector of genotype names with 2 or more simultaneous mutations, corresponding to PGs.

multi_muts

A list where each component corresponds to a genotype in multi_PGs, represented as a vector of gene names to indicate the mutations. mutations, corresponding to PGs.

germ

A matrix of numeric values used to modify the likelihood using germline testing, returned by germlineContrib, or 1.

marker

A matrix of numeric values used to modify the likelihood using marker testing, returned by markerContrib, or 1.

proband

A numeric value or vector of the unique IDs in ped corresponding to the proband(s).

max.mut

The maximum number of simultaneous mutations allowed.

af

A matrix of allele frequencies, where each row corresponds to the allele frequencies of a relative in ped.

net

A logical value indicating whether to return net or crude future risk estimates.

rr.bcrat

A numeric vector of length 2 containing the proband's BCRAT relative risks for ages < 50 and ages >= 50.

rr.pop

A data frame of race-specific for estimates of 1 / (1 - population attributable risk).

doclist

A list of density and survival arrays for death by other causes for the probands in proband, returned by calcCancerPenetrance. They will be used to calculate future risk.

age.by

The fixed age interval for the future risk calculation.

homo_idx

List of indices for genes with heterozygous and homozygous mutations. Each element consists of a vector with the index pair corresponding to the heterozygous and homozygous mutations for a given gene.

Value

A list with two components:

A matrix of posterior carrier probabilities returned by pp.peelingParing. The rows represent the relatives with IDs in proband and the columns represent the genotypes. If no genes were in the model specification, the values in the matrix are set to 1 for all probands and genotypes.

future.risk: A list where each element corresponds to a proband in proband and consists of a list of future risk estimates, as returned by calcFutureRisk. If no cancers were in the model specification, each proband's list of future risk estimates is replaced by the message "No future risk estimates were requested by the model specification."