Check pedigree structure and connectivity

Checks that all required columns in pedigree are present, valid, and consistent, printing informative error/warning messages and making corrections when possible. It also removes family members that are not connected to the proband(s) and splits the pedigree into a list of connected sub-families, if necessary. A list of relationships for each connected pedigree is also returned.

.pedStructureCheck(
  ped,
  cancers = .getCancersFromFam(ped),
  proband = NULL,
  unknown.race = UNKNOWN_RACE,
  unknown.ancestry = UNKNOWN_ANCESTRY,
  ignore.proband.germ = FALSE,
  allow.age.zero = FALSE,
  keep.only.model.cancers = TRUE
)

Arguments

ped

A pedigree data frame.

cancers

A vector of short cancer names corresponding to the cancers specified by the model that should be checked in ped. See PanelPRO:::CANCER_NAME_MAP$short for the list of options. The default is .getCancersFromFam(ped), which is a vector of all the short cancer names included in ped.

proband

A numeric value or vector of the unique IDs in ped corresponding to the proband(s). The default is NULL, in which case probands will be inferred from the isProband column in ped, if it exists and is valid. Otherwise, the IDs in proband will override the contents of ped$isProband. The user can also specify proband = "All", in which case all IDs in the ped will be treated as probands.

unknown.race

A character string indicating the default race to use when ped$race is missing or unsupported. The default is PanelPRO:::UNKNOWN_RACE.

unknown.ancestry

A character string indicating the default ancestry to use when ped$Ancestry is missing or unsupported. The default is PanelPRO:::UNKNOWN_ANCESTRY.

ignore.proband.germ

A logical value indicating if the proband(s)'s germline testing results (if provided in ped) should be considered. The default is FALSE, in which case the proband(s)'s germline testing results will not be ignored.

allow.age.zero

A logical value indicating whether zero ages can be reported in ped. The default is FALSE, in which case 0 values in age columns will be treated as missing. allow.age.zero should only be TRUE if individuals are confirmed to born with a cancer.

keep.only.model.cancers

A logical value indicating if only the cancers in cancers should be kept in ped. The default is TRUE.

Value

A list with three components:

ped_list: A list where each element is a connected pedigree that is a subset of the relatives in ped. Typically, ped_list has length 1, unless there are multiple probands who are related to multiple disconnected pedigrees. Relatives who are not connected to the proband(s) will be removed.

proband: A numeric value or vector of the unique IDs in the pedigree corresponding to the proband(s). If the proband argument was specified by the user, the returned list component will be identical. Otherwise, the IDs will correspond to individuals in the ped argument where ped$isProband == 1.

rel_ls: A list of the same length as ped_list, where each component is a list of relationships for each pedigree in ped_list.

Details

This function performs the following structural and connectivity checks, issuing informative error/warning messages and making corrections when possible:

Check expected columns

Check that all required columns (ID, Sex, MotherID, FatherID, CurAge, Twins, isDead, race, riskmod, interAge, Ancestry) are present in ped, and raise a ColumnNameNotMatch error if not.

Check probands

If a non-NULL proband argument was specified, the proband argument will override the contents of the ped$isProband column. If proband == "All", all IDs in the pedigree will be used. Otherwise, check if all IDs in proband are present in the pedigree, and raise a ProbandNotValid error if not. If the proband argument does not match with the probands implied by ped$isProband, print a ProbandOverwrite message informing the user that the probands in ped$isProband have been overridden by proband.

If the proband argument is NULL or unspecified by the user, search ped for the isProband column, and raise a ProbandNotValid error if it is not present in the data frame. Then, attempt to infer probands from the isProband column, treating NA values as 0 (not probands). If no probands (encoded as 1) are specified by isProband, raise a ProbandNotValid error. If isProband contains unexpected values, raise a TypeCheck error.

Raise a ProbandMissingAge error if any proband's CurAge is missing.

Raise a ProbandAgeOutOfBounds warning if any proband's CurAge is equal to PanelPRO:::MAXAGE, in which case their future risk will not be estimated.

Check cancers and ages

Columns in ped with the naming convention isAffX are used to encode cancer affection status and columns with the naming convention AgeX are used to encode affection ages. X should be one of the cancer short names in PanelPRO::CANCER_NAME_MAP$short.

Raise a CancerNotFound error if no isAffX columns can be found in ped.

Raise a BCNotExist error if CBC specified in the model but isAffBC or AgeBC columns cannot be found.

Raise a CBC_BCAge error if at least one subject has a BC age greater than their CBC Age.

Raise a CBC_Risk_Indicator error if codings of FirstBCType, AntiEstrogen, HRPreneoplasia, and/or BreastDensity columns have unrecognized values.

Raise a ModelCancersNotExist error if ped does not contain isAffX columns for all of the cancers in in the model specification cancers.

Raise an AffectionAgeNotFound error if any cancer does not have an AgeX column.

If the keep.only.model.cancers argument is set to TRUE (the default behavior), remove the isAffX and AgeX columns from ped corresponding to cancers that are not in cancers.

Values other than 0 and 1 in the isAffX columns will be recoded as -999, accompanied by an InvalidNA message.

If any individual is unaffected for a cancer but reports an affection age for the cancer, the affection age will be set to NA and a CancerAgeMisSpecified message will be raised.

If any individual develops cancer after their current age, a CancerAfterCurrentAge error will be raised along with a suggestion to substitute the maximum cancer age for the current age, if the user is unable to resolve the age inputs otherwise.

If the allow.age.zero argument is set to FALSE (the default behavior), set values in the AgeX and CurAge columns that are 0 to NA and print a FalseAgeZero message. These values will be treated as missing instead of 0. Values of AgeX and CurAge between 0 and 1 will be rounded up to 1; other decimal ages will be rounded to the nearest integer. In both cases, an AgeRounding message will be printed.

Check IDs and personal information

IDs should be unique identifiers in the pedigree. If any IDs in the ped$ID column are duplicated, raise a PedIDInvalid error.

Check that all IDs in the MotherID and FatherID columns can be found in ped$ID. Set parent IDs that cannot be found in the ID to be NA, so that they will be treated as missing from the pedigree, and raise an IDNotExist warning.

Set missing or unsupported values of ped$race and ped$Ancestry to the default race and ancestry specified by unknown.race and unknown.ancestry, and raise an InvalidNA message.

Set unexpected values of ped$Sex to NA and raise an InvalidInputForcingNA warning.

If ped$isDead is unknown (NA), assume that the individual is alive, set the value to 0, and raise an InvalidNA message.

Check that sexes of parents are valid, i.e. that all IDs in MotherID are female and all IDs in Father ID are male. Raise an InvalidParentSex error if this is not the case.

Check identical twins

If ped$Twins is unknown (NA), assume that the individual is not an identical twin, set the value to 0, and raise an InvalidNA message.

The following steps describe consistency checks among homozygotic twins and other identical multiple births. For more details, see the .checkTwins documentation.

If only one twin is present in a twin set encoded by ped$Twins, remove the individual's twin status by setting it to 0 and print a MissingTwinPair message

Assume that twins must have the same FatherID, MotherID, and Sex. If this information is missing for one (or more) twins but present in the other(s), infer the missing information from the individuals where it is known and print a MissingTwinInfoModified message. If the twins have unresolvable mismatched information, remove the twin label and raise a TwinInfoConflict warning.

Assume that living twins must have the same CurAge. If there is a disagreement in CurAge among living twins, the proband's CurAge overrides the others if one of the twins is a proband, and a TwinInfoConflict warning will be raised. Otherwise, the average CurAge will be assigned to all twins, and a TwinInfoConflict warning will be raised.

If the CurAge of a living twin is less than the CurAge of a dead twin, the average CurAge will be assigned to all twins, and a TwinInfoConflict warning will be raised.

Assume that twins must have the same race and Ancestry. If there is a disagreement, assign the default race and ancestry specified by unknown.race and unknown.ancestry to all twins, and raise a TwinInfoConflict warning.

Check risk modifiers/interventions

Individuals who report NA for riskmod and interAge are assumed to have no risk modifiers, and an InvalidInputForcingNone warning will be raised.

If an individual's riskmod and interAge are of different lengths, the longer will be truncated to match the length of the shorter. A RiskmodInconsistency message will be raised. Intervention ages between 0 and 1 will be rounded up to 1; other decimal ages will be rounded to the nearest integer. In both cases, an AgeRounding message will be printed.

Unsupported risk modifiers and invalid intervention ages will be removed and an InvalidInputForcingNone message will be printed.

Raise a DuplicatedRiskmod error if an individual has duplicated risk modifiers stored in riskmod.

Raise an InvalidRiskmod error if an individual's intervention is not preventative for the relevant cancer.

For individuals who do not report any risk modifiers in ped, set their value of riskmod to "None" and interAge to 1.

Check tumor biomarker testing

If an individual reports marker testing results in ped, but they are not affected for the related cancer, remove their marker testing results and raise an InvalidTests warning.

If ignore.proband.germ = TRUE, remove proband(s)'s germline testing information from ped.

Check germline testing

If germline testing results are reported in ped, print a GermlineTestAssumption message to inform the user that the default variant (see PanelPRO:::DEFAULT_VARIANTS) will be assumed for each mutation.

Typechecks

If any value of AgeX or CurAge exceeds PanelPRO:::MAXAGE, set the values to PanelPRO:::MAXAGE and raise an ExceedsMAXAGE warning.

Raise a TypeCheck error if Sex, Twins, or isAffX is missing (NA) for any individual.

Convert NA values to -999 and raise an InvalidNA message.

Check connectivity

Attempt to correct inconsistent hereditary features (race and Ancestry) among parents and children, and print a HeredityFix message. For more details, see the .checkHeredity documentation.

Remove family members who are disconnected from the proband(s), raising a DisconectedFamilyRemoval warning if so. For more details, see the .removeDisconnected documentation.

Resolve inconsistent sex information. If an individual's sex is incompatible with their parental identity, set the corresponding parent ID for their children to be missing. If an individual's sex is incompatible with their cancer affection status, set the corresponding isAffX value to 0 and raise a GenderCancerMismatch warning. For more details, see the .checkGender documentation. Finally, encode the Sex column as "Female"/"Male" instead of 0/1.

Validate the AgeX_lower and AgeX_upper columns for bounding cancer affection age, or create them if necessary. If a lower bound is greater than the corresponding upper bound, raise an AgeBounds error. For more details, see the .boundAffectionAges documentation.

Check that ped does not have any invalid mating patterns, i.e. intermarriages that create "loops", and raise a BadMating error if not. For more details, see the .checkMating documentation.

See also