Random number generation for a truncated normal distribution

Random number generation for a truncated normal distribution

.rtruncnorm(N, mean = 0, sd = 1, a = -Inf, b = Inf)

Arguments

N
The number of random observations to generate.

mean
Vector of means, to be passed into pnorm. The default is 0.

sd
Vector of standard deviations, to be passed into pnorm. The default is 1.

a
A numeric value; the lower truncation bound. The default is -Inf for no lower bound.

b
A numeric value; the upper truncation bound. The default is -Inf for no upper bound.

Value

A vector of N random numbers generated from a truncated normal distribution with mean mean, standard deviation sd, and bounds defined by a and b.

See also