Ensure that pedigree data frame column only contains allowed values

Used to check categorical and logical columns in checkFam.

forcingNA_ifnot_contains(
  ped,
  col_spec,
  allowed,
  default_NA_rep = NULL,
  NA_Ok = TRUE,
  vec = NULL,
  force_NA = TRUE,
  force_remove = FALSE
)

Arguments

ped

A pedigree data frame.

col_spec

The name of a column if ped with a specific attribute.

allowed

A vector of the values that are allowed in ped[[col_spec]].

default_NA_rep

The value to default to if NAs are encountered. The default is NULL.

NA_Ok

A logical value indicating whether NAs can be present in ped[[col_spec]]. The default is TRUE.

vec

A vector. The default is NULL. If vec is not NULL, the function will check vec instead of ped[[col_spec]].

force_NA

A logical value indicating whether non-permissible should be replaced with NAs. The default is TRUE.

force_remove

A logical value indicating whether non-permissible values should be removed. The default is FALSE.

Value

The vector ped[[col_spec]] (or vec if vec is not NULL) with non-permissible values either removed and replaced.