Format gene names

Convenience function for formatting gene names into shorter, more visually pleasing names. A message will be printed if choice of formatting results in non-unique gene names.

formatGeneNames(gene_names, format = "drop_hetero_anyPV")

Arguments

gene_names

A character vector of gene names, assumed to be of the format GENE_HETERO/HOMO_VARIANT, where GENE is the gene, HETERO/HOMO is either "hetero" or "homo" for the respective mutations, and VARIANT is either "anyPV" or a specific variant.

format

One of the following character strings:

"drop_hetero_anyPV" (default): Removes "hetero" and "anyPV" from gene names, but retains "homo" and specific variants.

"drop_hetero": Removes "hetero" from gene names, but retains "homo" and variant information.

"drop_anyPV": Removes "anyPV" from gene names, but retains "homo" and specific variants.

"drop_PVs": Removes variant information from gene names, but retains "hetero" and "homo".

"drop_heterohomo": Removes "hetero" and "homo" from gene names, but retains variant information.

"only_gene": Removes "hetero", "homo", and all variant information from gene names, retaining only the genes.

"full": Returns the gene names, unmodified. This is also the behavior if an unsupported format is specified.

Value

A character vector of formatted gene names.

Examples

gene_names = c("MUTYH_hetero_anyPV", "MUTYH_homo_anyPV", 
               "CHEK2_hetero_1100delC", 
               "MUTYH_hetero_anyPV.CHEK2_hetero_1100delC", 
               "MUTYH_homo_anyPV.CHEK2_hetero_1100delC")
PanelPRO:::formatGeneNames(gene_names)
#> [1] "MUTYH"                     "MUTYH_homo"               
#> [3] "CHEK2_1100delC"            "MUTYH.CHEK2_1100delC"     
#> [5] "MUTYH_homo.CHEK2_1100delC"
PanelPRO:::formatGeneNames(gene_names, format = "only_gene")
#> Choice of formatting results in non-unique gene names.
#> [1] "MUTYH"       "MUTYH"       "CHEK2"       "MUTYH.CHEK2" "MUTYH.CHEK2"