Get CBC model parameters for a subject.

These parameters can be used for subsetting a contralateral breast cancer penetrance array for carriers and/or creating noncarrier penetrances using CBCnoncarrierCrude.

getCBCParams(ped, id, cbc.penets, imputed.ages)

Arguments

ped

a checked pedigree data frame object returned by checkFam with at least columns ID, MotherID, FatherID, Sex, CurAge, isAffBC, AgeBC. Optional columns are race, FirstBCType, ER, AntiEstrogen, HRPreneoplasia, and BreastDensity.

id

numeric value, a subject ID in the pedigree for whom to find the results.

cbc.penets

CBC carrier penetrance array structured like PanelPRODatabase$Contralateral.

imputed.ages

list of imputed ages as returned by checkFam, must contain a AgeBC component.

Value

named list containing:

Gene: string, subject's genotype; either "noncarrier" or value from CBC_GENE_VARIANT_TYPES.

Sex: string, sex of the subject, either c("Female", "Male").

HadBC: binary indicator if subject has had a primary breast cancer yet.

FirstBCAge: numeric value, age of subject's first breast cancer.

FirstBCType: numeric value indicating whether the subject's first breast cancer was pure invasive or mixed invasive and DCIS, pure invasive or unknown type = 3 and mixed invasive and DCIS = 2. PanelPRO does not handle pure DCIS breast cancers.

FDRBC: numeric value indicating if the subject has first degree relatives with a history of breast cancer, Yes = 1 or No = 2.

ER: numeric value indicating subject's ER result, Unk = 3, Pos = 2, Neg = 1.

FirstBirth: numeric value indicating age category when subject's first child was born, <30/nulliparous = 1, 30-39 = 2, 40+ = 3, Unk = 4.

AntiEstrogen: numeric value indicating if anti-estrogen therapy was used to treat the 1st breast cancer, Yes = 1, No = 2, Unk = 3.

HRPreneoplasia: numeric value indicating a history of high risk preneoplasia, Yes = 1, No/Unk = 2.

BreastDensity: numeric value indicating BI-RADS breast density category. These codes are different from non-Hispanic Black women and everyone else as required by the cbcrisk package. For everyone else: "a" or entirely fat = 4, "b" or scattered = 3, "c" or heterogeneously dense = 2, "d" / extremely dense = 1, Unk = 5. For women who ARE non-Hispanic Black: "a" (entirely fat) or "b" (scattered) are 1, "c" (heterogeneously dense) or "d" (extremely dense) are 2, Unk is 3.

Details

If the subject has more than one positive CBC-linked gene, the gene selected will be the one with the highest CBC crude penetrance at age 85. If no germline testing information is available for the CBC genes, "noncarrier" status is assumed.

If age of first breast cancer has absent from the raw pedigree, a randomly selected imputed age from checkFam will be used instead.

The outputs of this function are designed to be compatible with the inputs to CBCnoncarrierCrude.