Coefficients in BCRAT relative risk model

Coefficients in BCRAT relative risk model

log.rr.bcrat

Format

Race-specific coefficients from Gail et al. (1989) for White women, Gail et al. (2007) for Black women, Matsuno et al. (2011) for Asian women, and Banegas et al. (2017) for Hispanic women, saved as a data frame with 6 rows and 7 columns:

beta1: Coefficient for number of benign breast biopsies.

beta2: Coefficient for age at menarche.

beta3: Coefficient for age at first live birth.

beta4: Coefficient for number of female first-degree relatives with breast cancer.

beta5: Coefficient for interaction between number of benign breast biopsies and age>=50.

beta6: Coefficient for interaction between age at first live birth and number of female first-degree relatives with breast cancer.

Race: "White", "Black", "Hispanic", "FHispanic", "Asian", or "Other"; "Hispanic" corresponds to U.S.-born Hispanic, while "FHispanic" corresponds to foreign-born Hispanic.

References

Gail MH, Brinton LA, Byar DP, Corle DK, Green SB, Schairer C, Mulvihill JJ. Projecting individualized probabilities of developing breast cancer for white females who are being examined annually. JNCI: Journal of the National Cancer Institute. 1989 Dec 20;81(24):1879-86.

Gail MH, Costantino JP, Pee D, Bondy M, Newman L, Selvan M, Anderson GL, Malone KE, Marchbanks PA, McCaskill-Stevens W, Norman SA. Projecting individualized absolute invasive breast cancer risk in African American women. Journal of the National Cancer Institute. 2007 Dec 5;99(23):1782-92.

Matsuno RK, Costantino JP, Ziegler RG, Anderson GL, Li H, Pee D, Gail MH. Projecting individualized absolute invasive breast cancer risk in Asian and Pacific Islander American women. Journal of the National Cancer Institute. 2011 Jun 22;103(12):951-61.

Banegas MP, John EM, Slattery ML, Gomez SL, Yu M, LaCroix AZ, Pee D, Chlebowski RT, Hines LM, Thompson CA, Gail MH. Projecting individualized absolute invasive breast cancer risk in US Hispanic women. Journal of the National Cancer Institute. 2017 Feb 1;109(2):djw215.